### 2025-07-09 #8558 @editorialbot
[PRE REVIEW]: search-query: An Open-Source Python Library for Academic Search Queries
**Submitting author:** @geritwagner (<a href="http://orcid.org/0000-0003-3926-7717">Gerit Wagner</a>)
**Repository:** https://github.com/CoLRev-Environment/search-query
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.12.0
