### 2025-07-09 734e39f
add missing DOI and update paper
