+++ paper.bib
@@ -16,6 +16,7 @@
    year                          = {2022},
    pages                         = {25--32},
    language                      = {eng},
+   doi                           = {10.1145/3548659.3561310},
 }
 
 @inproceedings{SinghSingh2017,
