### 2025-07-09 #8558 @geritwagner
Thank you for opening the pre-review issue. Based on the reference check summary, I added a missing DOI ([here](https://github.com/CoLRev-Environment/search-query/commit/734e39f87116d8cc0bc2301cb636479108d4f0c6)). The second item in the "missing DOI" section is a false positive; the suggested DOI points to a product review, not the original Yale MeSH Analyzer tool.
