### 2025-07-29 #8558 @geritwagner
Hi all,

Just to resurface this for consideration: As reviewers, the following have expertise relevant to the submission:

- TomDonoghue – Expertise in Python-based literature review tools; submitting author of LISC.
- ayuei – Expertise in information retrieval.
- CBID2 – Expertise in literature studies in the behavioral sciences.
- cliffckerr – Expertise in scientific software development in Python.

Additionally, DrMattG, timClicks, and linuxscout have previously reviewed literature review tools and may bring a useful perspective.

Thanks for considering!
