### 2025-08-06 192234a
Update paper.bib
