+++ paper.bib
@@ -308,5 +308,6 @@ author = {Wagner, Gerit and Prester, Julian},
 title = {{CoLRev}: An open-source environment for collaborative reviews},
 url = {https://github.com/CoLRev-Environment/colrev},
 version = {0.14.0},
-year = {2025}
+year = {2025},
+doi = {10.5281/zenodo.11668338},
 }
