### 2025-08-06 #8558 @geritwagner
Thank you for initiating the process — no worries at all about the timing. I reviewed the DOIs and added the one that was missing. The remaining entries refer to specific tools or Bachelor's theses, which do not have DOIs.
