### 2025-08-07 #8558 @CBID2
> @CBID2 @cliffckerr hello! any interest in reviewing this paper?

I'll give it a shot! :)
