### 2025-08-10 #8775 @editorialbot
[REVIEW]: search-query: An Open-Source Python Library for Academic Search Queries
**Submitting author:** @geritwagner (<a href="http://orcid.org/0000-0003-3926-7717">Gerit Wagner</a>)
**Repository:** https://github.com/CoLRev-Environment/search-query
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.15.0
