### 2025-08-10 CoLRev-Environment/search-query#48 @amcandio
JOSS review: installation guide / quickstart in README
This is a minor comment. The installation and quickstart section is great, would be a nicer experience if it is included as part of the README of the package, which is a common practice for libraries.

https://github.com/openjournals/joss-reviews/issues/8775
