### 2025-08-10 CoLRev-Environment/search-query#49 @amcandio
JOSS review: parse errors exit the program
When running into a parse query error:
```
from search_query.parser import parse

query_string = '("digital health"[Title/Abstract]) AND ("privacy"[Title/Abstract]'
query = parse(query_string, platform="pubmed")
# Output:
# ❌ Fatal: unbalanced-parentheses (PARSE_0002)
#   - Unbalanced opening parenthesis
#   Query: ("digital health"[Title/Abstract]) AND ("privacy"[Title/Abstract]
#                                                ^^^
```

This couples the command-line use cases with the library use case. In its current form, users are not given the chance to catch the error and handle it, prompting the whole program to exit. For example, if a system handling multiple requests depends on the library and receives a mal-formed query, the whole service would be down. This could lead to security concerns for users.

The common practice for this is to just raise an exception. Then command-line programs handle the exception and exit accordingly.

https://github.com/openjournals/joss-reviews/issues/8775
