### 2025-08-10 CoLRev-Environment/search-query#50 @amcandio
JOSS review: extend on some documentation section
Some sections such as https://colrev-environment.github.io/search-query/translate.html could benefit from some introductory paragraphs. This is great functionality but code is just presented without a clear introduction.
