### 2025-08-10 CoLRev-Environment/search-query#51 @amcandio
JOSS review [non-blocking]: consider unifying CLI into one command with sub-commands
**This is not a blocker for the paper.**

CLI functionality is great (https://colrev-environment.github.io/search-query/cli.html). However, it can harm usability to have multiple commands. A common pattern is to have one command and then have the actions as sub-commands. For example:

```
search-query lint
```

```
search-query translate
```

This would help unify interface and ease addition of new actions in the future.

https://github.com/openjournals/joss-reviews/issues/8775
