### 2025-08-10 CoLRev-Environment/search-query#52 @amcandio
JOSS review: community guidelines
This documentation is great https://colrev-environment.github.io/search-query/dev_docs/overview.html but I believe contributors guideline could be more inviting and more extense. Consider adding a CONTRIBUTING file to the package explaining how to build the source, how to add and run tests, etc.

This would be particularly interesting if the community starts helping adding support for new databases.

Regarding seeking support and reporting issues, README file should state what users should do when they have an issue, querion or they require support.

https://github.com/openjournals/joss-reviews/issues/8775
