### 2025-08-10 CoLRev-Environment/search-query#53 @amcandio
JOSS review: examples should be self-cointained
For example, example here https://colrev-environment.github.io/search-query/save.html depends on already having a query object. All examples should be self-contained and have all the relevant context to be executed.

https://github.com/openjournals/joss-reviews/issues/8775
