### 2025-08-10 CoLRev-Environment/search-query#54 @amcandio
JOSS review [non-blocking]: saving queries in structured format
**This is not a blocker for the paper.**

Storing queries as string can have issues with respect to (de)serialization changes and backwards compatibility.

For example, let's say one website decides to replace or "OR" in their queries by "|". All the old queries stored as string will become invalid. Once the parser is updated, they will all fail to parse, potentially needing rewrite. However, if you store them as a structured object such as a json (e.g. `{"operator": "OR", terms: [...]}`), then you don't need to change your stored queries once parsers change.

You can still store them as string for human-readability, but having both can have the benefit of robustness.

https://github.com/openjournals/joss-reviews/issues/8775
