### 2025-08-10 #8775 @amcandio
Really good work here! I added a bunch of issues on the repository, mostly related to documentation. Some of those are not blockers but more like general improvements for the package.

- https://github.com/CoLRev-Environment/search-query/issues/48
- https://github.com/CoLRev-Environment/search-query/issues/49
- https://github.com/CoLRev-Environment/search-query/issues/50
- https://github.com/CoLRev-Environment/search-query/issues/51
- https://github.com/CoLRev-Environment/search-query/issues/52
- https://github.com/CoLRev-Environment/search-query/issues/53
- https://github.com/CoLRev-Environment/search-query/issues/54
