### 2025-08-12 86c79f6
set up unified cli
