### 2025-08-13 CoLRev-Environment/search-query#54 @k-schnickmann
No worries! Just to clarify: the versioned parser approach I mentioned earlier is something we’re planning to implement, not something that’s already in place.

In the meantime, here’s a quick example of serializing a PubMed query into a generic search string:
```python
from search_query.parser import parse

query_string = '("digital health"[tiab]) AND ("privacy"[tiab])'
pubmed_query = parse(query_string, platform="pubmed")

generic_query = pubmed_query.translate(target_syntax="generic")
generic_query.to_generic_string()
# AND[OR["digital health"[abstract], "digital health"[title]], OR["privacy"[abstract], "privacy"[title]]]
```
We’d add something like this to the docs, along with instructions for manually saving both the platform-specific and generic strings in the search file. As mentioned, our idea is to eventually handle syntax changes through versioned parsers, rather than automatically storing both queries for every use case — as, in our view, that would introduce redundancy.
