### 2025-08-13 973420b
Unify CLI with sub-commands (#55)
* set up unified cli

* update platform handling

* update tests

* update docs
