### 2025-08-13 1dbe17c
raise exception instead of sys.exit()
