### 2025-08-13 aa74f1f
print linter success info
