### 2025-08-13 7bc65b6
Parser: raise exception instead of sys.exit() (#58)
* raise exception instead of sys.exit()

* remove pylint notes
