### 2025-08-13 CoLRev-Environment/search-query#49 @geritwagner
I revised the parser to raise an exception instead of `sys.exit()`. This should enable other systems to handle mal-formed queries gracefully.
