### 2025-08-13 CoLRev-Environment/search-query#48 @geritwagner
Agreed. We revised the README file accordingly.
