### 2025-08-14 4adec79
Community guidelines (#57)
* Add contribution guidelines

* format

* Update CONTRIBUTING.md

---------

Co-authored-by: Gerit Wagner <geritwagner@users.noreply.github.com>
