### 2025-08-14 CoLRev-Environment/search-query#52 @geritwagner
We’ve addressed this by adding a CONTRIBUTING.md with guidelines for running/extending tests and adding new platforms ([here](https://github.com/CoLRev-Environment/search-query/pull/57)), and updating the REAMDE.md ([here](https://github.com/CoLRev-Environment/search-query/pull/56)).
