### 2025-08-16 b524ee8
draft for pre_commit and translate description
