+++ docs/source/pre_commit.rst
@@ -3,6 +3,12 @@
 Pre-commit hooks
 ==========================================================
 
+The ``search-query-lint`` hook is a linter that automatically checks your 
+search query files (e.g., JSON queries) for errors and formatting issues.
+It is based on the ``search-query`` package and ensures that only valid,
+well-structured queries are committed to your repository. This helps catch
+mistakes early and keeps your project consistent without extra manual work.
+
 Linters can be included as pre-commit hooks by adding the following to the ``.pre-commit-config.yaml``:
 
 .. code-block:: yaml
