### 2025-08-16 02928f4
add examples to linter
