### 2025-08-16 259e6f4
Update pre_commit.rst
