### 2025-08-16 cc6f46b
Update translate.rst
