### 2025-08-16 f92804b
Update translate.rst
