### 2025-08-16 960dbd9
Update translate.rst
