### 2025-08-17 eb67270
add introduction to tranlate function
