### 2025-08-17 d84b584
add introduction to "improve-function" documentation
