### 2025-08-17 0c0d32c
extend "improve function" docs
