### 2025-08-17 fedec29
revert changes to translate docs
