+++ docs/source/translate.rst
@@ -3,13 +3,6 @@
 Translate
 ==========================================================
 
-The translate function translates a query from one database syntax to another. 
-Supported syntaxes are: Web of Science, PubMed, EBSCO and a generic syntax.
-If the syntax of the given query, or the requested target syntax, is from an unknown platform, the function will raise an exception.
-The output of the translate function is a query object.
-Additionally, using the to_string() function, the query object can be transformed to a string, ready to be used on the according platform. 
-
-
 ..
    TODO
    also describe how to translate to list format (flag/option for to-string methods)
