### 2025-08-17 9653c92
make example self-contained
