+++ docs/source/save.rst
@@ -8,6 +8,9 @@ To write a query to a JSON file, run the serializer:
 .. code-block:: python
 
     from search_query import SearchFile
+    from search_query.and_query import AndQuery
+
+    query = AndQuery(["digital", "work"], field="title")
 
     search_file = SearchFile(
         filename="search-file.json",
