### 2025-08-17 02df065
feat: extend documentation for "improve", make example self-contained (#60)
* add introduction to tranlate function

* add introduction to "improve-function" documentation

* extend "improve function" docs

* fix typo

* revert changes to translate docs

* make example self-contained
