### 2025-08-17 ed2fcb0
Added draft for description in translation and linter documentation (#59)
* draft for pre_commit and translate description

* add examples to linter

* fix one line

* Update pre_commit.rst

* Update translate.rst

* Update translate.rst

* Update translate.rst

---------

Co-authored-by: Gerit Wagner <geritwagner@users.noreply.github.com>
