+++ docs/source/dev_docs/_autosummary/search_query.cli.rst
@@ -13,5 +13,4 @@ search\_query.cli
 
    .. autosummary::
 
-      lint
-      translate
+      main
