### 2025-08-30 b9412d2
refactor(translator): inline platform translators into versioned modules
