+++ search_query/ebsco/v1_0_0/__init__.py
@@ -1,0 +1,4 @@
+"""EBSCOHost v1.0.0."""
+
+__author__ = """Gerit Wagner"""
+__email__ = "gerit.wagner@hec.ca"
