### 2025-09-01 4907872
remove redundant code in translators
