+++ search_query/pubmed/v1_0_0/translator.py
@@ -2,318 +2,10 @@
 """PubMed translator for version 1.0.0."""
 from __future__ import annotations
 
-import typing
-from collections import defaultdict
-from itertools import permutations
-
-from search_query.constants import Fields
-from search_query.constants import Operators
-from search_query.pubmed.constants import generic_field_to_syntax_field
-from search_query.pubmed.constants import syntax_str_to_generic_field_set
-from search_query.query import Query
-from search_query.query import SearchField
-from search_query.query_near import NEARQuery
-from search_query.query_or import OrQuery
-from search_query.query_term import Term
-from search_query.translator_base import QueryTranslator
+from search_query.pubmed.translator import PubmedTranslator
 
 
-class PubMedTranslator_v1_0_0(QueryTranslator):
+class PubMedTranslator_v1_0_0(PubmedTranslator):
     """Translator for Pubmed queries."""
 
     VERSION = "1.0.0"
-
-    @classmethod
-    def _expand_flat_or_chains(cls, query: Query) -> bool:
-        """Expand flat OR chains into a single OR query."""
-
-        if not (query.operator and query.value == Operators.OR):
-            return False
-        if not all(not child.operator for child in query.children):
-            return False
-        assert all(child.field for child in query.children)
-
-        fields = {child.field.value for child in query.children if child.field}
-
-        if len(fields) != 1:
-            return False
-
-        if next(iter(fields)) in {"[title and abstract]", "[tiab]"}:
-            existing_children = list(query.children)
-            for child in existing_children:
-                if not child.field:  # pragma: no cover
-                    continue
-                child.field.value = Fields.TITLE
-                new_child = Term(
-                    value=child.value,
-                    field=SearchField(value=Fields.ABSTRACT),
-                )
-                query.add_child(new_child)
-
-            return True
-
-        # elif ... other cases?
-
-        return False  # pragma: no cover
-
-    @classmethod
-    def _translate_fields(cls, query: Query) -> None:
-        if query.operator:
-            for child in query.children:
-                cls._translate_fields(child)
-
-        else:
-            if query.field and query.field.value not in ["[tiab]"]:
-                query.field.value = generic_field_to_syntax_field(query.field.value)
-
-    @classmethod
-    def _combine_tiab(cls, query: Query) -> None:
-        """Recursively combine identical terms from TI and AB into TIAB."""
-
-        if query.operator and query.value == "OR":
-            # ab does not exist: always expand to tiab
-            terms = []
-            for child in query.children:
-                if not child.operator and child.field and child.field.value == "ab":
-                    child.field.value = "[tiab]"
-                    terms.append(child.value)
-
-            if terms:
-                print(f"Info: combining terms from AB OR TI to TIAB: {terms}")
-
-            # Warn if the same terms are not available with ti
-            missing_terms = []
-            for term in terms:
-                if not any(
-                    term == child.value and child.field and child.field.value == "ti"
-                    for child in query.children
-                ):
-                    missing_terms.append(term)
-            if missing_terms:
-                print(
-                    "Info/Warning: Search field broadened for term "
-                    "(AB "
-                    "(without corresponding search for the same term with TI)"
-                    " -> TIAB): "
-                    f"{missing_terms}"
-                )
-
-            # Remove duplicates with ti
-            new_children = []
-            for child in query.children:
-                if child.operator:
-                    # unconditionally append operators
-                    new_children.append(child)
-                elif child.field and not (
-                    child.field.value == "ti" and child.value in terms
-                ):
-                    new_children.append(child)
-            query.children = new_children
-
-        # Recursively apply to child querys
-        for child in query.children:
-            cls._combine_tiab(child)
-
-    # pylint: disable=too-many-locals, too-many-branches
-    @classmethod
-    def _collapse_near_queries(cls, query: Query) -> Query:
[…]
