+++ search_query/translator.py
@@ -6,7 +6,7 @@ import typing
 
 from packaging.version import Version
 
-from search_query.ebsco.v1_0_0.translator import EBCOTranslator_v1_0_0
+from search_query.ebsco.v1_0_0.translator import EBSCOTranslator_v1_0_0
 from search_query.pubmed.v1_0_0.translator import PubMedTranslator_v1_0_0
 from search_query.translator_base import QueryTranslator
 from search_query.wos.v1_0_0.translator import WOSTranslator_v1_0_0
@@ -18,7 +18,7 @@ if typing.TYPE_CHECKING:
 TRANSLATORS: Dict[str, Dict[str, Type[QueryTranslator]]] = {
     "pubmed": {"1.0.0": PubMedTranslator_v1_0_0},
     "wos": {"1.0.0": WOSTranslator_v1_0_0},
-    "ebscohost": {"1.0.0": EBCOTranslator_v1_0_0},
+    "ebscohost": {"1.0.0": EBSCOTranslator_v1_0_0},
 }
 
 

[…]
