+++ test/pubmed/test_pubmed_parser_v_1_0_0.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-"""Tests for Pubmed search query parser."""
+"""Tests for PubMedParser_v1_0_0"""
 import pytest
 
 from search_query.constants import Colors
@@ -11,9 +11,8 @@ from search_query.exception import ListQuerySyntaxError
 from search_query.exception import SearchQueryException
 from search_query.parser import parse
 from search_query.pubmed.linter import PubmedQueryStringLinter
-from search_query.pubmed.parser import PubmedListParser
-from search_query.pubmed.parser import PubmedParser
-from search_query.pubmed.translator import PubmedTranslator
+from search_query.pubmed.v1_0_0.parser import PubMedListParser_v1_0_0
+from search_query.pubmed.v1_0_0.parser import PubMedParser_v1_0_0
 
 # to run (from top-level dir): pytest test/test_parser_pubmed.py
 
@@ -90,7 +89,7 @@ def test_tokenization(query_str: str, expected_tokens: list) -> None:
         f"Run query parser for: \n  {Colors.GREEN}{query_str}{Colors.END}\n--------------------\n"
     )
 
-    parser = PubmedParser(query_str)
+    parser = PubMedParser_v1_0_0(query_str)
     parser.tokenize()
     assert parser.tokens == expected_tokens, print(parser.tokens)
 
@@ -855,7 +854,7 @@ def test_linter(
         f"Run query parser for: \n  {Colors.GREEN}{query_str}{Colors.END}\n--------------------\n"
     )
 
-    parser = PubmedParser(query_str, field_general=field_general)
+    parser = PubMedParser_v1_0_0(query_str, field_general=field_general)
     try:
         parser.parse()
     except SearchQueryException:
@@ -984,7 +983,7 @@ def test_linter_with_general_field(
         f"Run query parser for: \n  {Colors.GREEN}{query_str}{Colors.END}\n--------------------\n"
     )
 
-    parser = PubmedParser(query_str, field_general=field_general)
+    parser = PubMedParser_v1_0_0(query_str, field_general=field_general)
     try:
         parser.parse()
     except SearchQueryException:
@@ -1053,7 +1052,7 @@ def test_parser(query_str: str, field_general: str, expected_parsed: str) -> Non
         f"Run query parser for: \n  {Colors.GREEN}{query_str}{Colors.END}\n--------------------\n"
     )
 
-    parser = PubmedParser(query_str, field_general=field_general)
+    parser = PubMedParser_v1_0_0(query_str, field_general=field_general)
     query = parser.parse()
 
     assert expected_parsed == query.to_generic_string(), print(
@@ -1068,7 +1067,7 @@ def test_list_parser_case_1() -> None:
 3. #1 OR #2
 """
 
-    list_parser = PubmedListParser(query_list=query_list)  #
+    list_parser = PubMedListParser_v1_0_0(query_list=query_list)  #
     list_parser.parse()
 
 
@@ -1078,7 +1077,7 @@ def test_list_parser_case_2() -> None:
 2. (acrobatics[Title/Abstract] OR aikido[Title/Abstract] OR archer[Title/Abstract] OR athletics[Title/Abstract])
 3. #1 AND #2 AND #4
 """
-    list_parser = PubmedListParser(query_list=query_list)
+    list_parser = PubMedListParser_v1_0_0(query_list=query_list)
     try:
         list_parser.parse()
     except ListQuerySyntaxError:
@@ -1106,7 +1105,7 @@ def test_list_parser_case_3() -> None:
 """
     print(query_list)
 
-    list_parser = PubmedListParser(query_list=query_list)
[…]
