### 2025-09-02 782a5c2
docs: document versioned components
