+++ docs/source/save.rst
@@ -16,9 +16,39 @@ To write a query to a JSON file, run the serializer:
         filename="search-file.json",
         query_str=query.to_string(platform="wos"),
         platform="wos",
+        version="1.0.0",
         authors=[{"name": "Tom Brady"}],
         record_info={},
         date={}
     )
 
     search_file.save()
+
+Backward compatibility
+----------------------
+
+Saved search strings include a ``version`` field so they can be
+re-parsed with the exact syntax they were created with:
+
+.. code-block:: json
+
+   {
+       "platform": "wos",
+       "search_string": "TS=(quantum dot)",
+       "version": "1.0.0"
+   }
+
+Queries may optionally be stored in a generic or structured form for
+redundancy:
+
+.. code-block:: json
+
+   {
+       "platform": "wos",
+       "search_string": "TS=(quantum dot)",
+       "version": "1.0.0",
+       "generic_query": "AND[quantum[TS=], dot[TS=]]"
+   }
+
+By default, search-query relies on the version pinning; the generic form
+is purely optional and mainly facilitates upgrades.
