+++ docs/source/dev_docs/parser_development.rst
@@ -1,6 +1,28 @@
 Parser
 ============
 
+Versioned namespaces
+--------------------
+
+Parsers live in versioned modules such as
+``search_query/pubmed/v1_0_0/parser.py``. Keeping previous versions
+allows reproducible parsing and backward compatibility.
+
+The central registry in ``search_query.parser`` exposes a ``PARSERS``
+mapping and resolves the appropriate version at runtime. Calling
+``parse(..., parser_version="latest")`` loads the highest available
+version for the chosen platform.
+
+Version policy follows ``MAJOR.MINOR.PATCH`` (semantic versioning):
+
+* **MAJOR** – breaking changes in syntax.
+* **MINOR** – backward compatible features.
+* **PATCH** – bug fixes or minor improvements.
+
+When introducing a new parser version, copy the previous versioned
+directory, adjust the implementation, and register the version in the
+``PARSERS`` dictionary.
+
 
 1. Inherit from Base Classes
 -----------------------------------
