### 2025-09-03 4d78b0e
update docs and wos: deprecated fields
