+++ search_query/wos/constants.py
@@ -153,8 +153,8 @@ field_GENERAL_TO_SYNTAX_MAP = {
     "Year Published": "PY=",
     "Affiliation": "AD=",
     "Funding Agency": "FO=",
-    "Publisher": "PU=",
-    "Publiation Date": "PD=",
+    # "Publisher": "PU=", # deprecated ?
+    # "Publiation Date": "PY=", # deprecated ?
     "Abstract": "AB=",
     "Accession Number": "UT=",
     "Address": "AD=",
