+++ docs/source/dev_docs/_autosummary/search_query.pubmed.v1_0_0.parser.rst
@@ -1,0 +1,21 @@
+search\_query.pubmed.v1\_0\_0.parser
+====================================
+
+.. automodule:: search_query.pubmed.v1_0_0.parser
+
+
+
+
+
+
+
+
+
+
+
+   .. rubric:: Classes
+
+   .. autosummary::
+
+      PubMedListParser_v1_0_0
+      PubMedParser_v1_0_0
