+++ docs/source/dev_docs/_autosummary/search_query.wos.v1_0_0.rst
@@ -1,0 +1,32 @@
+search\_query.wos.v1\_0\_0
+==========================
+
+.. automodule:: search_query.wos.v1_0_0
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :recursive:
+
+   search_query.wos.v1_0_0.parser
+   search_query.wos.v1_0_0.serializer
+   search_query.wos.v1_0_0.translator
