+++ docs/source/dev_docs/_autosummary/search_query.ebsco.v1_0_0.serializer.rst
@@ -1,0 +1,20 @@
+search\_query.ebsco.v1\_0\_0.serializer
+=======================================
+
+.. automodule:: search_query.ebsco.v1_0_0.serializer
+
+
+
+
+
+
+
+
+
+
+
+   .. rubric:: Classes
+
+   .. autosummary::
+
+      EBCOSerializer_v1_0_0
