+++ docs/source/dev_docs/_autosummary/search_query.generic.serializer.rst
@@ -9,8 +9,12 @@ search\_query.generic.serializer
 
 
 
-   .. rubric:: Functions
+
+
+
+
+   .. rubric:: Classes
 
    .. autosummary::
 
-      to_string_generic
+      GenericSerializer

[…]
