### 2025-09-04 56f125b
add docs on upgrade and versioned translators
