+++ docs/source/index.rst
@@ -144,7 +144,7 @@ Below is a high-level overview of the core functionalities:
    :caption: Platforms
 
    platforms/platform_index
-   platforms/upgrade
+   platforms/syntax_upgrade
 
 .. toctree::
    :hidden:
