### 2025-09-04 754b5cd
feat: add dynamic registry discovery
