+++ search_query/registry.py
@@ -1,0 +1,178 @@
+from __future__ import annotations
+
+import importlib
+import re
+import typing as _t
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import Dict, Type
+
+from packaging.version import Version
+
+from search_query.constants import PLATFORM
+from search_query.parser_base import QueryListParser, QueryStringParser
+from search_query.serializer_base import StringSerializer as QueryStringSerializer
+from search_query.serializer_base import StringSerializer as QueryListSerializer
+from search_query.translator_base import QueryTranslator
+
+_VERSION_DIR_RE = re.compile(r"^v_(\d+)_(\d+)_(\d+)$")
+_MODULE_FILES = ("parser", "serializer", "translator")
+
+def _to_version_str(dir_name: str) -> str:
+    m = _VERSION_DIR_RE.match(dir_name)
+    if not m:
+        raise ValueError(f"Invalid version directory name: {dir_name!r}")
+    return ".".join(m.groups())
+
+def _platform_pkg_path(platform: str) -> Path:
+    pkg = f"search_query.{platform}"
+    mod = importlib.import_module(pkg)
+    return Path(next(iter(getattr(mod, "__path__"))))
+
+@dataclass
+class _Registry:
+    # Parsers
+    parsers: Dict[str, Dict[str, Type[QueryStringParser]]] = field(default_factory=dict)
+    list_parsers: Dict[str, Dict[str, Type[QueryListParser]]] = field(default_factory=dict)
+    # Serializers
+    serializers: Dict[str, Dict[str, Type[QueryStringSerializer]]] = field(default_factory=dict)
+    list_serializers: Dict[str, Dict[str, Type[QueryListSerializer]]] = field(default_factory=dict)
+    # Translators
+    translators: Dict[str, Dict[str, Type[QueryTranslator]]] = field(default_factory=dict)
+
+    # --- register methods (fail on duplicates) ---
+    def register_parser_string(self, platform: str, version: str, cls: Type[QueryStringParser]) -> None:
+        self.parsers.setdefault(platform, {})
+        if version in self.parsers[platform]:
+            prev = self.parsers[platform][version]
+            raise RuntimeError(f"Duplicate string parser for {platform} {version}: {prev.__name__} vs {cls.__name__}")
+        self.parsers[platform][version] = cls  # type: ignore[assignment]
+
+    def register_parser_list(self, platform: str, version: str, cls: Type[QueryListParser]) -> None:
+        self.list_parsers.setdefault(platform, {})
+        if version in self.list_parsers[platform]:
+            prev = self.list_parsers[platform][version]
+            raise RuntimeError(f"Duplicate list parser for {platform} {version}: {prev.__name__} vs {cls.__name__}")
+        self.list_parsers[platform][version] = cls  # type: ignore[assignment]
+
+    def register_serializer_string(self, platform: str, version: str, cls: Type[QueryStringSerializer]) -> None:
+        self.serializers.setdefault(platform, {})
+        if version in self.serializers[platform]:
+            prev = self.serializers[platform][version]
+            raise RuntimeError(f"Duplicate string serializer for {platform} {version}: {prev.__name__} vs {cls.__name__}")
+        self.serializers[platform][version] = cls  # type: ignore[assignment]
+
+    def register_serializer_list(self, platform: str, version: str, cls: Type[QueryListSerializer]) -> None:
[…]
