+++ search_query/wos/v_1_0_0/parser.py
@@ -22,3 +22,8 @@ class WOSListParser_v1_0_0(WOSListParser):
         super().__init__(query_list=query_list, field_general=field_general)
         self.parser_class = WOSParser_v1_0_0
         self.linter = WOSQueryListLinter(self, WOSParser_v1_0_0)
+
+
+def register(registry, *, platform: str, version: str) -> None:
+    registry.register_parser_string(platform, version, WOSParser_v1_0_0)
+    registry.register_parser_list(platform, version, WOSListParser_v1_0_0)
