+++ search_query/ebscohost/v_1_0_0/serializer.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python3
+"""EBSCO serializer for version 1.0.0."""
+from __future__ import annotations
+
+from search_query.ebscohost.serializer import EBSCOQuerySerializer
+
+# pylint: disable=too-few-public-methods
+
+
+class EBCOSerializer_v1_0_0(EBSCOQuerySerializer):
+    """EBSCO serializer for version 1.0.0."""
+
+    VERSION = "1.0.0"
+
+
+def register(registry, *, platform: str, version: str) -> None:
+    registry.register_serializer_string(platform, version, EBCOSerializer_v1_0_0)
+    registry.register_serializer_list(platform, version, EBCOSerializer_v1_0_0)
