### 2025-09-05 9cff523
rename Registry class
