+++ search_query/ebscohost/v_1_0_0/parser.py
@@ -2,9 +2,13 @@
 """Versioned EBSCO parser wrappers."""
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from search_query.ebscohost.linter import EBSCOListLinter
-from search_query.ebscohost.parser import EBSCOListParser
-from search_query.ebscohost.parser import EBSCOParser
+from search_query.ebscohost.parser import EBSCOListParser, EBSCOParser
+
+if TYPE_CHECKING:  # pragma: no cover
+    from search_query.registry import Registry
 
 # pylint: disable=too-few-public-methods
 
@@ -26,6 +30,8 @@ class EBSCOListParser_v1_0_0(EBSCOListParser):
         self.linter = EBSCOListLinter(self, EBSCOParser_v1_0_0)
 
 
-def register(registry, *, platform: str, version: str) -> None:
+def register(registry: Registry, *, platform: str, version: str) -> None:
+    """Register these parsers with the ``registry``."""
+
     registry.register_parser_string(platform, version, EBSCOParser_v1_0_0)
     registry.register_parser_list(platform, version, EBSCOListParser_v1_0_0)
