+++ search_query/wos/v_1_0_0/translator.py
@@ -2,7 +2,12 @@
 """Web of Science translator for version 1.0.0."""
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from search_query.wos.translator import WOSTranslator
+
+if TYPE_CHECKING:  # pragma: no cover
+    from search_query.registry import Registry
 
 
 class WOSTranslator_v1_0_0(WOSTranslator):
@@ -11,5 +16,7 @@ class WOSTranslator_v1_0_0(WOSTranslator):
     VERSION = "1.0.0"
 
 
-def register(registry, *, platform: str, version: str) -> None:
+def register(registry: Registry, *, platform: str, version: str) -> None:
+    """Register this translator with the ``registry``."""
+
     registry.register_translator(platform, version, WOSTranslator_v1_0_0)

[…]
