+++ search_query/pubmed/v_1_0_0/serializer.py
@@ -2,18 +2,23 @@
 """PubMed serializer for version 1.0.0."""
 from __future__ import annotations
 
+import typing
+
 from search_query.pubmed.serializer import PUBMEDQuerySerializer
+
+if typing.TYPE_CHECKING:  # pragma: no cover
+    from search_query.registry import Registry
 
 
 # pylint: disable=too-few-public-methods
-
-
 class PubMedSerializer_v1_0_0(PUBMEDQuerySerializer):
     """PubMed serializer for version 1.0.0."""
 
     VERSION = "1.0.0"
 
 
-def register(registry, *, platform: str, version: str) -> None:
+def register(registry: Registry, *, platform: str, version: str) -> None:
+    """Register this serializer with the ``registry``."""
+
     registry.register_serializer_string(platform, version, PubMedSerializer_v1_0_0)
     registry.register_serializer_list(platform, version, PubMedSerializer_v1_0_0)
