+++ search_query/wos/v_1_0_0/__init__.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 """WoS v1.0.0."""
 
 __author__ = """Gerit Wagner"""
