+++ search_query/ebscohost/v_1_0_0/__init__.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 """EBSCOHost v1.0.0."""
 
 __author__ = """Gerit Wagner"""
