### 2025-09-08 c5395b4
revise docs on syntax upgrades
