+++ docs/source/platforms/syntax_upgrade.rst
@@ -4,40 +4,98 @@ Syntax Upgrades
 ===============
 
 The ``search-query`` package allows you to **upgrade database-specific queries**
-from one syntax version to another.
+from one *database syntax version* to another.
 
 Most users will interact with this feature via the **command-line interface (CLI)**.
-Behind the scenes, the upgrade follows a structured procedure using a generic query, or
-intermediate representation (IR).
+Behind the scenes, the upgrade follows a structured procedure using a generic query
+as an intermediate representation (IR).
 
 Quick Start (CLI)
 -----------------
 
-To upgrade a query stored in ``query.json`` to version ``2.0.0``:
+To upgrade a query stored in ``query.json`` to version ``2.0.0`` of a given database syntax:
 
 .. code-block:: console
 
    search-query upgrade query.json --to 2.0.0
 
 If ``--to`` is omitted or set to ``latest``, the tool automatically selects
-the latest supported version for the specified platform.
+the latest syntax version for the specified database.
+The upgraded query will be written back to the file (or another location if
+specified with ``--output``).
 
-.. tip::
+.. note::
 
-   The upgraded query will be written back to the file (or another location if
-   specified with ``--output``).
+   **What does "syntax version" mean?**
+
+   A *syntax version* identifies the **evolving query language of a database platform**
+   (e.g., PubMed, Web of Science), not the version of this package.
+
+   The latest package version (of ``search-query``) evolves independently and
+   contains the most up-to-date parser/translator/serializer for the syntax versions it supports
+   across databases.
+
+Key Concepts
+------------
+
+- **Database Syntax Version (DSV)**
+  The identifier of the query language as implemented by a database provider
+  (e.g., supported operators, search fields, precedence rules).
+
+- **Package Platform Version**
+  The version of the local ``search-query`` package/connector.
+  It may increment independently from database syntax versions.
+  New package releases can improve or extend support for *existing* DSVs
+  without implying that the database syntax itself changed.
+
+- **Observability**
+  Discoverability and testability of a DSV in the wild.
+  Observability of *historic* versions is often limited because providers rarely
+  document all changes systematically.
+  What matters most is that the **current/latest DSV is observable** (i.e., can be tested)
+  and that **material evolutions** in a database’s syntax can be recognized.
 
 Versioning Policy
-----------------------------
+-----------------
 
-Version policy follows ``MAJOR.MINOR.PATCH`` (semantic versioning):
+Version identifiers follow a pragmatic schema aligned with semantic versioning,
+with emphasis on **MAJOR** and **MINOR** to reflect real-world database syntax evolution:
 
-* **MAJOR** – breaking changes in syntax.
-* **MINOR** – backward compatible features.
-* **PATCH** – bug fixes or minor improvements.
+* **MAJOR** – *Breaking* syntax changes in the **database**
+  (e.g., operators changed/removed, search fields renamed/removed, precedence rules altered).
+  Older queries may need migration.
+
+* **MINOR** – *Backward-compatible* syntax extensions
+  (e.g., new search fields, additional operators, increased limits on terms or clauses).
+  Older queries still parse and run.
+
 
 ..
-   This structure allows for limited observability of syntax changes, focusing on current and future versions.
+   * **PATCH** – Non-semantic tweaks in this package’s handling of a DSV
+   (bug fixes, robustness, diagnostics). Patch bumps do **not** imply a database syntax change.
+
+.. important::
+
+   The **primary use cases** are to
+
+   1. store queries in the **current DSV**,
+   2. **upgrade** them to a newer DSV when available, and
+   3. **anticipate** that the database syntax may evolve (occasionally via new MAJOR versions).
+
+   Full support for arbitrarily many **historic** DSVs
[…]
