+++ docs/source/dev_docs/_autosummary/search_query.ebscohost.v_1_0_0.translator.rst
@@ -1,0 +1,26 @@
+search\_query.ebscohost.v\_1\_0\_0.translator
+=============================================
+
+.. automodule:: search_query.ebscohost.v_1_0_0.translator
+
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      register
+
+
+
+
+
+   .. rubric:: Classes
+
+   .. autosummary::
+
+      EBSCOTranslator_v1_0_0
