+++ docs/source/dev_docs/_autosummary/search_query.ebscohost.rst
@@ -1,0 +1,35 @@
+search\_query.ebscohost
+=======================
+
+.. automodule:: search_query.ebscohost
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :recursive:
+
+   search_query.ebscohost.constants
+   search_query.ebscohost.linter
+   search_query.ebscohost.parser
+   search_query.ebscohost.serializer
+   search_query.ebscohost.translator
+   search_query.ebscohost.v_1_0_0
