+++ docs/source/dev_docs/_autosummary/search_query.ebsco.v1_0_0.translator.rst
@@ -1,20 +1,0 @@
-search\_query.ebsco.v1\_0\_0.translator
-=======================================
-
-.. automodule:: search_query.ebsco.v1_0_0.translator
-
-
-
-
-
-
-
-
-
-
-
-   .. rubric:: Classes
-
-   .. autosummary::
-
-      EBSCOTranslator_v1_0_0
