+++ docs/source/dev_docs/_autosummary/search_query.translator.rst
@@ -1,4 +1,0 @@
-search\_query.translator
-========================
-
-.. automodule:: search_query.translator
