+++ docs/source/dev_docs/_autosummary/search_query.ebsco.v1_0_0.rst
@@ -1,32 +1,0 @@
-search\_query.ebsco.v1\_0\_0
-============================
-
-.. automodule:: search_query.ebsco.v1_0_0
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-.. rubric:: Modules
-
-.. autosummary::
-   :toctree:
-   :recursive:
-
-   search_query.ebsco.v1_0_0.parser
-   search_query.ebsco.v1_0_0.serializer
-   search_query.ebsco.v1_0_0.translator
