+++ docs/source/dev_docs/_autosummary/search_query.ebsco.rst
@@ -1,35 +1,0 @@
-search\_query.ebsco
-===================
-
-.. automodule:: search_query.ebsco
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-.. rubric:: Modules
-
-.. autosummary::
-   :toctree:
-   :recursive:
-
-   search_query.ebsco.constants
-   search_query.ebsco.linter
-   search_query.ebsco.parser
-   search_query.ebsco.serializer
-   search_query.ebsco.translator
-   search_query.ebsco.v1_0_0
