+++ docs/source/dev_docs/_autosummary/search_query.wos.v1_0_0.translator.rst
@@ -1,20 +1,0 @@
-search\_query.wos.v1\_0\_0.translator
-=====================================
-
-.. automodule:: search_query.wos.v1_0_0.translator
-
-
-
-
-
-
-
-
-
-
-
-   .. rubric:: Classes
-
-   .. autosummary::
-
-      WOSTranslator_v1_0_0
