+++ docs/source/dev_docs/_autosummary/search_query.ebsco.v1_0_0.parser.rst
@@ -1,21 +1,0 @@
-search\_query.ebsco.v1\_0\_0.parser
-===================================
-
-.. automodule:: search_query.ebsco.v1_0_0.parser
-
-
-
-
-
-
-
-
-
-
-
-   .. rubric:: Classes
-
-   .. autosummary::
-
-      EBSCOListParser_v1_0_0
-      EBSCOParser_v1_0_0
