+++ docs/source/dev_docs/_autosummary/search_query.wos.v1_0_0.serializer.rst
@@ -1,20 +1,0 @@
-search\_query.wos.v1\_0\_0.serializer
-=====================================
-
-.. automodule:: search_query.wos.v1_0_0.serializer
-
-
-
-
-
-
-
-
-
-
-
-   .. rubric:: Classes
-
-   .. autosummary::
-
-      WOSSerializer_v1_0_0
