+++ docs/source/dev_docs/_autosummary/search_query.wos.v_1_0_0.parser.rst
@@ -1,0 +1,27 @@
+search\_query.wos.v\_1\_0\_0.parser
+===================================
+
+.. automodule:: search_query.wos.v_1_0_0.parser
+
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      register
+
+
+
+
+
+   .. rubric:: Classes
+
+   .. autosummary::
+
+      WOSListParser_v1_0_0
+      WOSParser_v1_0_0
