### 2025-09-08 58d937a
feat(wos): add version 0.0.0 supporting deprecated fields
