### 2025-09-08 80f8524
feat(wos): add version 0.0.0 supporting deprecated fields
