### 2025-09-08 3eab4ef
add note to docs
