### 2025-09-09 d4c8afa
test: verify upgrade from wos 0.0.0
