+++ search_query/__version__.py
@@ -5,4 +5,7 @@ except ImportError:  # pragma: no cover
     # For Python < 3.8
     from importlib_metadata import version  # type: ignore
 
-__version__ = version("search_query")
+try:
+    __version__ = version("search_query")
+except Exception:  # pragma: no cover - fallback when package metadata missing
+    __version__ = "0.0.0"
