### 2025-09-09 246b9f9
wos:0.0.0 translator and test
