+++ docs/source/dev_docs/_autosummary/search_query.wos.v_0_0_0.translator.rst
@@ -1,0 +1,35 @@
+search\_query.wos.v\_0\_0\_0.translator
+=======================================
+
+.. automodule:: search_query.wos.v_0_0_0.translator
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      register
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      WOSTranslator_v0_0_0
+   
+   
+
+   
+   
+   
+
+
+
