+++ docs/source/dev_docs/_autosummary/search_query.wos.rst
@@ -32,4 +32,5 @@ search\_query.wos
    search_query.wos.parser
    search_query.wos.serializer
    search_query.wos.translator
+   search_query.wos.v_0_0_0
    search_query.wos.v_1_0_0
