+++ docs/source/dev_docs/_autosummary/search_query.wos.v_0_0_0.rst
@@ -1,0 +1,33 @@
+search\_query.wos.v\_0\_0\_0
+============================
+
+.. automodule:: search_query.wos.v_0_0_0
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :recursive:
+
+   search_query.wos.v_0_0_0.parser
+   search_query.wos.v_0_0_0.serializer
+   search_query.wos.v_0_0_0.translator
+
