### 2025-09-09 5e2615b
update docs: syntax upgrade
