+++ docs/source/platforms/syntax_upgrade.rst
@@ -13,16 +13,20 @@ as an intermediate representation (IR).
 Quick Start (CLI)
 -----------------
 
-To upgrade a query stored in ``query.json`` to version ``2.0.0`` of a given database syntax:
+To upgrade a query stored in ``query.json`` to the latest version of a given database syntax:
 
 .. code-block:: console
 
-   search-query upgrade query.json --to 2.0.0
+   search-query upgrade query.json
 
-If ``--to`` is omitted or set to ``latest``, the tool automatically selects
-the latest syntax version for the specified database.
-The upgraded query will be written back to the file (or another location if
-specified with ``--output``).
+By default, if you omit the ``--to`` option or set it to ``latest``,
+the tool upgrades the query to the latest supported syntax version for the specified database.
+You can also specify a particular target version explicitly using ``--to <version>``,
+but in most cases, upgrading to the latest version (by omitting ``--to``) is recommended.
+The upgraded query will be written back to the file (or another location if specified with ``--output``).
+
+Versioning Policy
+-----------------
 
 .. note::
 
@@ -30,36 +34,14 @@ specified with ``--output``).
 
    A *syntax version* identifies the **evolving query language of a database platform**
    (e.g., PubMed, Web of Science), not the version of this package.
-
    The latest package version (of ``search-query``) evolves independently and
    contains the most up-to-date parser/translator/serializer for the syntax versions it supports
    across databases.
-
-Key Concepts
-------------
-
-- **Database Syntax Version (DSV)**
-  The identifier of the query language as implemented by a database provider
-  (e.g., supported operators, search fields, precedence rules).
-
-- **Package Platform Version**
-  The version of the local ``search-query`` package/connector.
-  It may increment independently from database syntax versions.
-  New package releases can improve or extend support for *existing* DSVs
-  without implying that the database syntax itself changed.
-
-- **Observability**
-  Discoverability and testability of a DSV in the wild.
-  Observability of *historic* versions is often limited because providers rarely
-  document all changes systematically.
-  What matters most is that the **current/latest DSV is observable** (i.e., can be tested)
-  and that **material evolutions** in a database’s syntax can be recognized.
-
-Versioning Policy
------------------
+   New package releases can improve or extend support for *existing* syntax versions
+   without implying that the database syntax itself changed.
 
 Version identifiers follow a pragmatic schema aligned with semantic versioning,
-with emphasis on **MAJOR** and **MINOR** to reflect real-world database syntax evolution:
+with emphasis on **MAJOR** and **MINOR** to reflect database syntax evolution:
 
 * **MAJOR** – *Breaking* syntax changes in the **database**
   (e.g., operators changed/removed, search fields renamed/removed, precedence rules altered).
@@ -69,7 +51,6 @@ with emphasis on **MAJOR** and **MINOR** to reflect real-world database syntax e
   (e.g., new search fields, additional operators, increased limits on terms or clauses).
   Older queries still parse and run.
 
-
 ..
    * **PATCH** – Non-semantic tweaks in this package’s handling of a DSV
    (bug fixes, robustness, diagnostics). Patch bumps do **not** imply a database syntax change.
@@ -78,23 +59,27 @@ with emphasis on **MAJOR** and **MINOR** to reflect real-world database syntax e
 
    The **primary use cases** are to
 
-   1. store queries in the **current DSV**,
-   2. **upgrade** them to a newer DSV when available, and
+   1. store queries in the **current syntax version**,
+   2. **upgrade** them to a newer syntax version when available, and
    3. **anticipate** that the database syntax may evolve (occasionally via new MAJOR versions).
 
-   Full support for arbitrarily many **historic** DSVs
+   Full support for arbitrarily many **historic** syntax versions
[…]
