+++ docs/source/dev_docs/_autosummary/search_query.wos.v_0_0_0.serializer.rst
@@ -3,33 +3,24 @@ search\_query.wos.v\_0\_0\_0.serializer
 
 .. automodule:: search_query.wos.v_0_0_0.serializer
 
-   
-   
-   
 
-   
-   
+
+
+
+
+
    .. rubric:: Functions
 
    .. autosummary::
-   
-      register
-   
-   
 
-   
-   
+      register
+
+
+
+
+
    .. rubric:: Classes
 
    .. autosummary::
-   
+
       WOSSerializer_v0_0_0
-   
-   
-
-   
-   
-   
-
-
-
