+++ docs/source/dev_docs/_autosummary/search_query.wos.v_0_0_0.rst
@@ -3,21 +3,21 @@ search\_query.wos.v\_0\_0\_0
 
 .. automodule:: search_query.wos.v_0_0_0
 
-   
-   
-   
 
-   
-   
-   
 
-   
-   
-   
 
-   
-   
-   
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 
@@ -30,4 +30,3 @@ search\_query.wos.v\_0\_0\_0
    search_query.wos.v_0_0_0.parser
    search_query.wos.v_0_0_0.serializer
    search_query.wos.v_0_0_0.translator
-
