+++ docs/source/dev_docs/_autosummary/search_query.wos.v_0_0_0.parser.rst
@@ -3,35 +3,26 @@ search\_query.wos.v\_0\_0\_0.parser
 
 .. automodule:: search_query.wos.v_0_0_0.parser
 
-   
-   
-   
 
-   
-   
+
+
+
+
+
    .. rubric:: Functions
 
    .. autosummary::
-   
-      register
-   
-   
 
-   
-   
+      register
+
+
+
+
+
    .. rubric:: Classes
 
    .. autosummary::
-   
+
       WOSListParser_v0_0_0
       WOSParser_v0_0_0
       WOSQueryStringLinter_v0_0_0
-   
-   
-
-   
-   
-   
-
-
-
