+++ docs/source/dev_docs/_autosummary/search_query.wos.v_0_0_0.translator.rst
@@ -3,33 +3,24 @@ search\_query.wos.v\_0\_0\_0.translator
 
 .. automodule:: search_query.wos.v_0_0_0.translator
 
-   
-   
-   
 
-   
-   
+
+
+
+
+
    .. rubric:: Functions
 
    .. autosummary::
-   
-      register
-   
-   
 
-   
-   
+      register
+
+
+
+
+
    .. rubric:: Classes
 
    .. autosummary::
-   
+
       WOSTranslator_v0_0_0
-   
-   
-
-   
-   
-   
-
-
-
