### 2025-09-09 6ac5489
feat: allow linters to ignore failures
