+++ search_query/ebscohost/v_1_0_0/parser.py
@@ -25,10 +25,24 @@ class EBSCOListParser_v1_0_0(EBSCOListParser):
 
     VERSION = "1.0.0"
 
-    def __init__(self, query_list: str, *, field_general: str = "") -> None:
-        super().__init__(query_list=query_list, field_general=field_general)
+    def __init__(
+        self,
+        query_list: str,
+        *,
+        field_general: str = "",
+        ignore_failing_linter: bool = False,
+    ) -> None:
+        super().__init__(
+            query_list=query_list,
+            field_general=field_general,
+            ignore_failing_linter=ignore_failing_linter,
+        )
         self.parser_class = EBSCOParser_v1_0_0
-        self.linter = EBSCOListLinter(self, EBSCOParser_v1_0_0)
+        self.linter = EBSCOListLinter(
+            parser=self,
+            string_parser_class=EBSCOParser_v1_0_0,
+            ignore_failing_linter=ignore_failing_linter,
+        )
 
 
 def register(registry: Registry, *, platform: str, version: str) -> None:
