+++ search_query/pubmed/v_1_0_0/parser.py
@@ -23,11 +23,25 @@ class PubMedListParser_v1_0_0(PubmedListParser):
 
     VERSION = "1.0.0"
 
-    def __init__(self, query_list: str, *, field_general: str = "") -> None:
-        super().__init__(query_list=query_list, field_general=field_general)
+    def __init__(
+        self,
+        query_list: str,
+        *,
+        field_general: str = "",
+        ignore_failing_linter: bool = False,
+    ) -> None:
+        super().__init__(
+            query_list=query_list,
+            field_general=field_general,
+            ignore_failing_linter=ignore_failing_linter,
+        )
         # Ensure versioned parser and linter are used
         self.parser_class = PubMedParser_v1_0_0
-        self.linter = PubmedQueryListLinter(self, PubMedParser_v1_0_0)
+        self.linter = PubmedQueryListLinter(
+            self,
+            PubMedParser_v1_0_0,
+            ignore_failing_linter=ignore_failing_linter,
+        )
 
 
 def register(registry: Registry, *, platform: str, version: str) -> None:
