### 2025-09-09 32e0b0f
prevent parser/translator failures with new linters
