+++ search_query/pubmed/parser.py
@@ -52,6 +52,7 @@ class PubmedParser(QueryStringParser):
         offset: typing.Optional[dict] = None,
         original_str: typing.Optional[str] = None,
         silent: bool = False,
+        ignore_failing_linter: bool = False,
     ) -> None:
         """Initialize the parser."""
         super().__init__(
@@ -59,9 +60,14 @@ class PubmedParser(QueryStringParser):
             field_general=field_general,
             offset=offset,
             original_str=original_str,
+            silent=silent,
+            ignore_failing_linter=ignore_failing_linter,
         )
         self.linter = PubmedQueryStringLinter(
-            query_str=query_str, original_str=original_str, silent=silent
+            query_str=query_str,
+            original_str=original_str,
+            silent=silent,
+            ignore_failing_linter=ignore_failing_linter,
         )
 
     def tokenize(self) -> None:
@@ -297,13 +303,19 @@ class PubmedListParser(QueryListParser):
         query_list: str,
         *,
         field_general: str = "",
+        ignore_failing_linter: bool = False,
     ) -> None:
         super().__init__(
             query_list,
             parser_class=PubmedParser,
             field_general=field_general,
+            ignore_failing_linter=ignore_failing_linter,
         )
-        self.linter = PubmedQueryListLinter(self, PubmedParser)
+        self.linter = PubmedQueryListLinter(
+            self,
+            PubmedParser,
+            ignore_failing_linter=ignore_failing_linter,
+        )
 
     def parse(self) -> Query:
         """Parse the query in list format."""
@@ -320,6 +332,7 @@ class PubmedListParser(QueryListParser):
             field_general=self.field_general,
             offset=offset,
             silent=True,
+            ignore_failing_linter=self.ignore_failing_linter,
         )
         try:
             query = query_parser.parse()
