+++ search_query/generic/linter.py
@@ -28,8 +28,15 @@ class GenericLinter(QueryStringLinter):
 
     VALID_fieldS_REGEX = re.compile(r"\b(?:" + "|".join(sorted(field_codes)) + r")\b")
 
-    def __init__(self, query_str: str = "") -> None:
-        super().__init__(query_str=query_str)
+    def __init__(
+        self,
+        query_str: str = "",
+        *,
+        ignore_failing_linter: bool = False,
+    ) -> None:
+        super().__init__(
+            query_str=query_str, ignore_failing_linter=ignore_failing_linter
+        )
 
     def syntax_str_to_generic_field_set(self, field_value: str) -> set:
         """Translate a search field"""
