+++ search_query/ebscohost/parser.py
@@ -55,6 +55,7 @@ class EBSCOParser(QueryStringParser):
         offset: typing.Optional[dict] = None,
         original_str: typing.Optional[str] = None,
         silent: bool = False,
+        ignore_failing_linter: bool = False,
     ) -> None:
         """Initialize the parser."""
         super().__init__(
@@ -62,9 +63,14 @@ class EBSCOParser(QueryStringParser):
             field_general=field_general,
             offset=offset,
             original_str=original_str,
+            silent=silent,
+            ignore_failing_linter=ignore_failing_linter,
         )
         self.linter = EBSCOQueryStringLinter(
-            query_str=query_str, original_str=original_str, silent=silent
+            query_str=query_str,
+            original_str=original_str,
+            silent=silent,
+            ignore_failing_linter=ignore_failing_linter,
         )
 
     def combine_subsequent_tokens(self) -> None:
@@ -428,13 +434,19 @@ class EBSCOListParser(QueryListParser):
         self,
         query_list: str,
         field_general: str = "",
+        ignore_failing_linter: bool = False,
     ) -> None:
         super().__init__(
             query_list=query_list,
             parser_class=EBSCOParser,
             field_general=field_general,
+            ignore_failing_linter=ignore_failing_linter,
         )
-        self.linter = EBSCOListLinter(parser=self, string_parser_class=EBSCOParser)
+        self.linter = EBSCOListLinter(
+            parser=self,
+            string_parser_class=EBSCOParser,
+            ignore_failing_linter=ignore_failing_linter,
+        )
 
     def parse(self) -> Query:
         """Parse EBSCO list query."""
@@ -455,6 +467,7 @@ class EBSCOListParser(QueryListParser):
             field_general=self.field_general,
             offset=offset,
             silent=True,
+            ignore_failing_linter=self.ignore_failing_linter,
         )
         try:
             query = query_parser.parse()
