+++ search_query/ebscohost/linter.py
@@ -67,8 +67,14 @@ class EBSCOQueryStringLinter(QueryStringLinter):
         *,
         original_str: typing.Optional[str] = None,
         silent: bool = False,
+        ignore_failing_linter: bool = False,
     ) -> None:
-        super().__init__(query_str=query_str, original_str=original_str, silent=silent)
+        super().__init__(
+            query_str=query_str,
+            original_str=original_str,
+            silent=silent,
+            ignore_failing_linter=ignore_failing_linter,
+        )
 
     def validate_tokens(
         self,
@@ -363,10 +369,15 @@ class EBSCOListLinter(QueryListLinter):
         self,
         parser: EBSCOListParser,
         string_parser_class: typing.Type[QueryStringParser],
-    ):
+        ignore_failing_linter: bool = False,
+    ) -> None:
         self.parser: EBSCOListParser = parser
         self.string_parser_class = string_parser_class
-        super().__init__(parser, string_parser_class)
+        super().__init__(
+            parser,
+            string_parser_class,
+            ignore_failing_linter=ignore_failing_linter,
+        )
 
     def validate_tokens(self) -> None:
         """Validate token list"""
