### 2025-09-09 e981c55
refactor: simplify version naming to major
