+++ test/pubmed/test_pubmed_parser_v_1_0_0.py
@@ -1,1195 +1,0 @@
-#!/usr/bin/env python
-"""Tests for PubMedParser_v1_0_0"""
-import pytest
-
-from search_query.constants import Colors
-from search_query.constants import PLATFORM
-from search_query.constants import QueryErrorCode
-from search_query.constants import Token
-from search_query.constants import TokenTypes
-from search_query.exception import ListQuerySyntaxError
-from search_query.exception import SearchQueryException
-from search_query.parser import parse
-from search_query.pubmed.linter import PubmedQueryStringLinter
-from search_query.pubmed.v_1_0_0.parser import PubMedListParser_v1_0_0
-from search_query.pubmed.v_1_0_0.parser import PubMedParser_v1_0_0
-
-# to run (from top-level dir): pytest test/test_parser_pubmed.py
-
-# flake8: noqa: E501
-
-
-@pytest.mark.parametrize(
-    "query_str, expected_tokens",
-    [
-        (
-            '("health tracking" [tw] OR"remote monitoring"[tw])AND wearable device[tw]NOT Comment[pt]',
-            [
-                Token(value="(", type=TokenTypes.PARENTHESIS_OPEN, position=(0, 1)),
-                Token(
-                    value='"health tracking"',
-                    type=TokenTypes.TERM,
-                    position=(1, 18),
-                ),
-                Token(value="[tw]", type=TokenTypes.FIELD, position=(19, 23)),
-                Token(value="OR", type=TokenTypes.LOGIC_OPERATOR, position=(24, 26)),
-                Token(
-                    value='"remote monitoring"',
-                    type=TokenTypes.TERM,
-                    position=(26, 45),
-                ),
-                Token(value="[tw]", type=TokenTypes.FIELD, position=(45, 49)),
-                Token(value=")", type=TokenTypes.PARENTHESIS_CLOSED, position=(49, 50)),
-                Token(value="AND", type=TokenTypes.LOGIC_OPERATOR, position=(50, 53)),
-                Token(
-                    value="wearable device",
-                    type=TokenTypes.TERM,
-                    position=(54, 69),
-                ),
-                Token(value="[tw]", type=TokenTypes.FIELD, position=(69, 73)),
-                Token(value="NOT", type=TokenTypes.LOGIC_OPERATOR, position=(73, 76)),
-                Token(value="Comment", type=TokenTypes.TERM, position=(77, 84)),
-                Token(value="[pt]", type=TokenTypes.FIELD, position=(84, 88)),
-            ],
-        ),
-        (
-            "home care workers [Title] OR health care workers [Title]",
-            [
-                Token(
-                    value="home care workers",
-                    type=TokenTypes.TERM,
-                    position=(0, 17),
-                ),
-                Token(value="[Title]", type=TokenTypes.FIELD, position=(18, 25)),
-                Token(value="OR", type=TokenTypes.LOGIC_OPERATOR, position=(26, 28)),
-                Token(
-                    value="health care workers",
-                    type=TokenTypes.TERM,
-                    position=(29, 48),
-                ),
-                Token(value="[Title]", type=TokenTypes.FIELD, position=(49, 56)),
-            ],
-        ),
-        (
-            'technician*[ Title/Abstract] OR "Personnel"[ MeSH Terms]',
-            [
-                Token(value="technician*", type=TokenTypes.TERM, position=(0, 11)),
-                Token(
-                    value="[ Title/Abstract]", type=TokenTypes.FIELD, position=(11, 28)
-                ),
-                Token(value="OR", type=TokenTypes.LOGIC_OPERATOR, position=(29, 31)),
[…]
