### 2025-09-09 4ebc7c2
Rename WOS v0 module to major version scheme and fix CLI tests
