+++ search_query/wos/v_1/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+"""WoS v1."""
+
+__author__ = """Gerit Wagner"""
+__email__ = "gerit.wagner@hec.ca"
