+++ test/wos/test_wos_parser_v_1_0_0.py
@@ -1,1246 +1,0 @@
-#!/usr/bin/env python3
-"""Tests for WOSParser_v1_0_0"""
-import typing
-
-import pytest
-
-from search_query.constants import Colors
-from search_query.constants import GENERAL_ERROR_POSITION
-from search_query.constants import PLATFORM
-from search_query.constants import Token
-from search_query.constants import TokenTypes
-from search_query.exception import ListQuerySyntaxError
-from search_query.exception import SearchQueryException
-from search_query.query import SearchField
-from search_query.query_and import AndQuery
-from search_query.query_or import OrQuery
-from search_query.query_term import Term
-from search_query.wos.v_1_0_0.parser import WOSListParser_v1_0_0
-from search_query.wos.v_1_0_0.parser import WOSParser_v1_0_0
-
-# ruff: noqa: E501
-# flake8: noqa: E501
-
-
-@pytest.mark.parametrize(
-    "query_str, expected_tokens",
-    [
-        (
-            "TI=example AND AU=John Doe",
-            [
-                Token(value="TI=", type=TokenTypes.FIELD, position=(0, 3)),
-                Token(value="example", type=TokenTypes.TERM, position=(3, 10)),
-                Token(value="AND", type=TokenTypes.LOGIC_OPERATOR, position=(11, 14)),
-                Token(value="AU=", type=TokenTypes.FIELD, position=(15, 18)),
-                Token(value="John Doe", type=TokenTypes.TERM, position=(18, 26)),
-            ],
-        ),
-        (
-            "TI=example example2 AND AU=John Doe",
-            [
-                Token(value="TI=", type=TokenTypes.FIELD, position=(0, 3)),
-                Token(
-                    value="example example2",
-                    type=TokenTypes.TERM,
-                    position=(3, 19),
-                ),
-                Token(value="AND", type=TokenTypes.LOGIC_OPERATOR, position=(20, 23)),
-                Token(value="AU=", type=TokenTypes.FIELD, position=(24, 27)),
-                Token(value="John Doe", type=TokenTypes.TERM, position=(27, 35)),
-            ],
-        ),
-        (
-            "TI=example example2 example3 AND AU=John Doe",
-            [
-                Token(value="TI=", type=TokenTypes.FIELD, position=(0, 3)),
-                Token(
-                    value="example example2 example3",
-                    type=TokenTypes.TERM,
-                    position=(3, 28),
-                ),
-                Token(value="AND", type=TokenTypes.LOGIC_OPERATOR, position=(29, 32)),
-                Token(value="AU=", type=TokenTypes.FIELD, position=(33, 36)),
-                Token(value="John Doe", type=TokenTypes.TERM, position=(36, 44)),
-            ],
-        ),
-        (
-            "TI=ex$mple* AND AU=John?Doe",
-            [
-                Token(value="TI=", type=TokenTypes.FIELD, position=(0, 3)),
-                Token(value="ex$mple*", type=TokenTypes.TERM, position=(3, 11)),
-                Token(value="AND", type=TokenTypes.LOGIC_OPERATOR, position=(12, 15)),
-                Token(value="AU=", type=TokenTypes.FIELD, position=(16, 19)),
-                Token(value="John?Doe", type=TokenTypes.TERM, position=(19, 27)),
-            ],
-        ),
-        (
-            'direction*OR (route AND test)"',
-            [
-                Token(value="direction*", type=TokenTypes.TERM, position=(0, 10)),
-                Token(value="OR", type=TokenTypes.LOGIC_OPERATOR, position=(10, 12)),
-                Token(value="(", type=TokenTypes.PARENTHESIS_OPEN, position=(13, 14)),
[…]
