+++ search_query/ebscohost/parser.py
@@ -18,6 +18,8 @@ from search_query.query import SearchField
 from search_query.query_near import NEARQuery
 from search_query.query_term import Term
 
+# pylint: disable=duplicate-code
+
 
 class EBSCOParser(QueryStringParser):
     """Parser for EBSCO queries."""
