### 2025-09-09 17e4d85
docs: document platform versions
