### 2025-09-09 317e348
docs: document platform versions (#72)
