+++ docs/source/platforms/platform_index.rst
@@ -5,15 +5,15 @@ Platforms
 
 Overview of supported platforms:
 
-+----------------+---------------------------+
-| Platform ID    | Documentation             |
-+================+===========================+
-| `wos`          | :doc:`wos <wos>`          |
-+----------------+---------------------------+
-| `ebscohost`    | :doc:`ebsco <ebsco>`      |
-+----------------+---------------------------+
-| `pubmed`       | :doc:`pubmed <pubmed>`    |
-+----------------+---------------------------+
++----------------+---------------------------+----------------+---------------------+
+| Platform ID    | Documentation             | Latest version | Previous versions   |
++================+===========================+================+=====================+
+| `wos`          | :doc:`wos <wos>`          | 1              | 0                   |
++----------------+---------------------------+----------------+---------------------+
+| `ebscohost`    | :doc:`ebsco <ebsco>`      | 1              | -                   |
++----------------+---------------------------+----------------+---------------------+
+| `pubmed`       | :doc:`pubmed <pubmed>`    | 1              | -                   |
++----------------+---------------------------+----------------+---------------------+
 
 .. toctree::
    :maxdepth: 1
