+++ docs/source/dev_docs/_autosummary/search_query.wos.v_1.serializer.rst
@@ -1,13 +1,23 @@
 search\_query.wos.v\_1.serializer
-=======================================
+=================================
 
 .. automodule:: search_query.wos.v_1.serializer
+
+
+
+
+
+
 
    .. rubric:: Functions
 
    .. autosummary::
 
       register
+
+
+
+
 
    .. rubric:: Classes
 
