+++ docs/source/dev_docs/_autosummary/search_query.pubmed.v_1.translator.rst
@@ -1,13 +1,23 @@
 search\_query.pubmed.v\_1.translator
-==========================================
+====================================
 
 .. automodule:: search_query.pubmed.v_1.translator
+
+
+
+
+
+
 
    .. rubric:: Functions
 
    .. autosummary::
 
       register
+
+
+
+
 
    .. rubric:: Classes
 
