+++ docs/source/dev_docs/_autosummary/search_query.ebscohost.v_1.parser.rst
@@ -1,13 +1,23 @@
 search\_query.ebscohost.v\_1.parser
-=========================================
+===================================
 
 .. automodule:: search_query.ebscohost.v_1.parser
+
+
+
+
+
+
 
    .. rubric:: Functions
 
    .. autosummary::
 
       register
+
+
+
+
 
    .. rubric:: Classes
 
