+++ docs/source/dev_docs/_autosummary/search_query.wos.v_0.rst
@@ -1,5 +1,5 @@
-search\_query.wos.v\_0\_0\_0
-============================
+search\_query.wos.v\_0
+======================
 
 .. automodule:: search_query.wos.v_0
 
