+++ docs/source/dev_docs/_autosummary/search_query.generic.v_1.serializer.rst
@@ -1,13 +1,23 @@
 search\_query.generic.v\_1.serializer
-===========================================
+=====================================
 
 .. automodule:: search_query.generic.v_1.serializer
+
+
+
+
+
+
 
    .. rubric:: Functions
 
    .. autosummary::
 
       register
+
+
+
+
 
    .. rubric:: Classes
 
