+++ docs/source/dev_docs/_autosummary/search_query.wos.v_1.parser.rst
@@ -1,10 +1,23 @@
+search\_query.wos.v\_1.parser
+=============================
+
 .. automodule:: search_query.wos.v_1.parser
+
+
+
+
+
+
 
    .. rubric:: Functions
 
    .. autosummary::
 
       register
+
+
+
+
 
    .. rubric:: Classes
 
