+++ docs/source/dev_docs/_autosummary/search_query.ebscohost.v_1.serializer.rst
@@ -1,13 +1,23 @@
 search\_query.ebscohost.v\_1.serializer
-=============================================
+=======================================
 
 .. automodule:: search_query.ebscohost.v_1.serializer
+
+
+
+
+
+
 
    .. rubric:: Functions
 
    .. autosummary::
 
       register
+
+
+
+
 
    .. rubric:: Classes
 
