+++ docs/source/dev_docs/_autosummary/search_query.wos.v_1.rst
@@ -1,4 +1,25 @@
+search\_query.wos.v\_1
+======================
+
 .. automodule:: search_query.wos.v_1
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 .. rubric:: Modules
 
