### 2025-09-09 b1d6946
Versioned parsers/serializers/translators + central registry + upgrades through generic queries (IR) (#62)
* add initial upgrade module

* implement first version

* remove redundant code in translators

* tests: versions of parsers

* update docs and wos: deprecated fields

* add docs on upgrade and versioned translators

* shorter heading

* update docs

* rename to ebscohost

* registry

* refactor

* pylint: cyclic import

* revise docs on syntax upgrades

* update docs

* wos:0.0.0

* add note to docs

* wos:0.0.0 translator and test

* update docs

* update docs: syntax upgrade

* docs: 
[…]
