+++ search_query/ebscohost/parser.py
@@ -1,0 +1,485 @@
+#!/usr/bin/env python3
+"""EBSCO query parser."""
+from __future__ import annotations
+
+import re
+import typing
+
+from search_query.constants import PLATFORM
+from search_query.constants import Token
+from search_query.constants import TokenTypes
+from search_query.ebscohost.linter import EBSCOListLinter
+from search_query.ebscohost.linter import EBSCOQueryStringLinter
+from search_query.exception import QuerySyntaxError
+from search_query.parser_base import QueryListParser
+from search_query.parser_base import QueryStringParser
+from search_query.query import Query
+from search_query.query import SearchField
+from search_query.query_near import NEARQuery
+from search_query.query_term import Term
+
+# pylint: disable=duplicate-code
+
+
+class EBSCOParser(QueryStringParser):
+    """Parser for EBSCO queries."""
+
+    PARENTHESIS_REGEX = re.compile(r"[\(\)]")
+    LOGIC_OPERATOR_REGEX = re.compile(r"\b(AND|OR|NOT)\b", flags=re.IGNORECASE)
+    PROXIMITY_OPERATOR_REGEX = re.compile(
+        r"(N|W)\d+|(NEAR|WITHIN)/\d+", flags=re.IGNORECASE
+    )
+    FIELD_REGEX = re.compile(r"\b([A-Z]{2})\b")
+    TERM_REGEX = re.compile(r"\"[^\"]*\"|\*?\b[^()\s]+")
+
+    OPERATOR_REGEX = re.compile(
+        "|".join([LOGIC_OPERATOR_REGEX.pattern, PROXIMITY_OPERATOR_REGEX.pattern])
+    )
+
+    pattern = re.compile(
+        "|".join(
+            [
+                PARENTHESIS_REGEX.pattern,
+                LOGIC_OPERATOR_REGEX.pattern,
+                PROXIMITY_OPERATOR_REGEX.pattern,
+                FIELD_REGEX.pattern,
+                TERM_REGEX.pattern,
+            ]
+        )
+    )
+
+    # pylint: disable=too-many-arguments
+    def __init__(
+        self,
+        query_str: str,
+        *,
+        field_general: str = "",
+        offset: typing.Optional[dict] = None,
+        original_str: typing.Optional[str] = None,
+        silent: bool = False,
+        ignore_failing_linter: bool = False,
+    ) -> None:
+        """Initialize the parser."""
+        super().__init__(
+            query_str,
+            field_general=field_general,
+            offset=offset,
+            original_str=original_str,
+            silent=silent,
+            ignore_failing_linter=ignore_failing_linter,
+        )
+        self.linter = EBSCOQueryStringLinter(
+            query_str=query_str,
+            original_str=original_str,
+            silent=silent,
+            ignore_failing_linter=ignore_failing_linter,
+        )
+
+    def combine_subsequent_tokens(self) -> None:
+        """Combine subsequent tokens based on specific conditions."""
+
+        combined_tokens = []
+        i = 0
+
+        while i < len(self.tokens):
+            # Iterate through token list
+            # current_token, current_token_type, position = self.tokens[i]
+
+            if self.tokens[i].type == TokenTypes.TERM:
+                # Filter out TERM
+                start_pos = self.tokens[i].position[0]
+                end_position = self.tokens[i].position[1]
+                combined_value = self.tokens[i].value
+
+                while (
+                    i + 1 < len(self.tokens)
+                    and self.tokens[i + 1].type == TokenTypes.TERM
+                ):
+                    # Iterate over subsequent terms and combine
+                    combined_value += f" {self.tokens[i + 1].value}"
+                    end_position = self.tokens[i + 1].position[1]
+                    i += 1
+
+                combined_tokens.append(
+                    Token(
[…]
