### 2025-09-09 2f29bf0
docs: example for saving structured search string
