### 2025-09-09 4bc2c34
docs: adopt impersonal tone
