+++ search_query/query_near.py
@@ -30,7 +30,7 @@ class NEARQuery(Query):
         platform: str = "generic",
     ) -> None:
         """init method
-        search terms: strings which you want to include in the search query
+        search terms: strings to include in the search query
         nested queries: queries whose roots are appended to the query
         nearDistance: distance of operator e.g. NEAR/2 --> near_distance = 2
         search field: search field to which the query should be applied
