+++ docs/source/platforms/pubmed.rst
@@ -14,7 +14,7 @@ Queries can be entered using either:
 - the `basic search box <https://pubmed.ncbi.nlm.nih.gov/>`_, or
 - the `advanced Query builder <https://pubmed.ncbi.nlm.nih.gov/advanced/>`_.
 
-These interfaces are functionally equivalent for the purposes of `search-query`. We recommend users always **copy and store the full query string** from either interface.
+These interfaces are functionally equivalent for the purposes of `search-query`. The full query string should always be **copied and stored** from either interface.
 
 .. tip::
 
@@ -31,7 +31,7 @@ When storing a PubMed query in a `.json` file or as a string:
 List Query Format
 --------------------
 
-PubMed allows users to **combine previous searches** using the search history panel. These list-based queries can reference earlier searches using numbered identifiers (e.g., `#1 OR #2`). In the advanced search interface, such combinations can also be created via the **“Add”** button. List queries are supported by `search-query` parsers.
+PubMed allows combining previous searches using the search history panel. These list-based queries can reference earlier searches using numbered identifiers (e.g., `#1 OR #2`). In the advanced search interface, such combinations can also be created via the **“Add”** button. List queries are supported by `search-query` parsers.
 
 List queries should be formatted as follows:
 
@@ -47,7 +47,7 @@ Best Practices and Recommendations
 
 The advanced PubMed interface offers a dropdown for search fields such as `[Title/Abstract]`, `[Author]`, etc. When using this feature:
 
-- If you apply the **same field** to the entire query, it is safe to store this value in the `general_field`.
+- When the **same field** applies to the entire query, storing this value in the `general_field` is safe.
 - If the query contains **multiple search fields**, `search-query` will treat each term individually, and **leaving `general_field` empty is preferred**.
 
 .. note::
