+++ docs/source/platforms/wos.rst
@@ -13,7 +13,7 @@ Queries in Web of Science can be constructed using:
 - the `Basic search interface <https://www.webofscience.com/wos/woscc/basic-search>`_, or
 - the `Advanced search interface <https://www.webofscience.com/wos/woscc/advanced-search>`_, which allows for precise control using field tags (e.g., `TI=`, `TS=`, `AU=`).
 
-When working with `search-query`, we recommend copying the **Advanced search query string** directly for use as the `search_string`.
+When working with `search-query`, the **Advanced search query string** should be copied directly for use as the `search_string`.
 
 Web of Science supports structured queries using field tags such as `TS=` (Topic), `TI=` (Title), `AB=` (Abstract), `AU=` (Author), etc.
 
@@ -37,7 +37,7 @@ Example::
 List Query Format
 ---------------------
 
-Web of Science allows users to **construct complex queries** by combining previously defined search sets using numbered references (e.g., ``#1 AND #2``). This list-based approach is commonly used in systematic searches where multiple search lines are logically combined.
+Web of Science allows the construction of **complex queries** by combining previously defined search sets using numbered references (e.g., ``#1 AND #2``). This list-based approach is commonly used in systematic searches where multiple search lines are logically combined.
 
 Such list queries are supported by the ``search-query`` parsers and allow referencing earlier statements using ``#`` followed by the search line number.
 
