+++ docs/source/dev_docs/overview.rst
@@ -35,7 +35,7 @@ Versioned implementations live inside
 To add a new parser or serializer version:
 
 1. copy the latest versioned directory (e.g. ``v1`` → ``v2``),
-2. apply your changes,
+2. apply the changes,
 3. register the new version in ``search_query.parser.PARSERS``,
    ``search_query.serializers.SERIALIZERS`` or
    ``search_query.translators.TRANSLATORS`` so it becomes discoverable.

[…]
