### 2025-09-09 d7a5a29
update docs for consistency
closes #50
