+++ CONTRIBUTING.md
@@ -1,26 +1,26 @@
 ## How to contribute
 
-Thank you for your interest in contributing to the *search-query* package!
-We encourage you to have a look at our [open issues](https://github.com/CoLRev-Environment/search-query/issues) or open a new one if you encounter an un-addressed problem.
+Contributions to the *search-query* package are welcome.
+The [open issues](https://github.com/CoLRev-Environment/search-query/issues) highlight outstanding problems and potential enhancements.
 
-Here are some guidelines for how to contribute to the package.
+The following guidelines describe how to contribute to the package.
 
 **Want to implement a new search platform?**
 
-Right now, the package supports a limited amount of search platforms.
-We therefore appreciate help in expanding the package with new platforms.
-If you’re unsure which platform to implement, check our [platform roadmap](https://github.com/CoLRev-Environment/search-query/issues/46).
+Right now, the package supports a limited number of search platforms.
+Help in expanding the package with new platforms is appreciated.
+The [platform roadmap](https://github.com/CoLRev-Environment/search-query/issues/46) indicates priority areas.
 
-Our [development documentation](https://colrev-environment.github.io/search-query/dev_docs/overview.html) includes an overview of best practices for implementing a new search platform, along with code skeletons to help you get started.
+The [development documentation](https://colrev-environment.github.io/search-query/dev_docs/overview.html) provides an overview of best practices for implementing a new search platform, along with code skeletons.
 
 **Found a bug?**
 
-If you found a bug or encountered any issues while using the package, you can contribute by [opening an issue](https://github.com/CoLRev-Environment/search-query/issues/new).
-When possible, include a minimal code example and describe the expected behavior.
+Bug reports and other issues can be submitted via [the issue tracker](https://github.com/CoLRev-Environment/search-query/issues/new).
+A minimal code example and a description of expected behavior facilitate triage.
 
 **Running and extending the test suite**
 
-We use [`pytest`](https://docs.pytest.org/) for testing.
+[`pytest`](https://docs.pytest.org/) is used for testing.
 Tests are located in the `tests/` directory and are organized by functionality (e.g., linting, parsing, platform-specific implementations).
 
 To run all tests locally:
@@ -52,15 +52,14 @@ Adding new tests
 * Group related tests into logical classes or functions.
 * Use clear and descriptive test names (`test_what_it_does`).
 * Prefer [pytest parameterization](https://docs.pytest.org/en/stable/how-to/parametrize.html) when testing multiple input–output pairs.
-* If you implement a new search platform, include tests for end-to-end search-query functionality
+* Implementing a new search platform should include tests for end-to-end search-query functionality.
 
 **Other ways to contribute**
 
-Of course, we also welcome smaller contributions, such as bug fixes or improvements to the package documentation.
+Smaller contributions, such as bug fixes or improvements to the package documentation, are also appreciated.
 <!-- TODO: Include guide for contributing to docs? -->
 
-If you’ve made changes to the source code or documentation, fork the repository and open a [pull request](https://github.com/CoLRev-Environment/search-query/compare).
-Please include a clear description of your changes.
+After making changes to the source code or documentation, fork the repository and open a [pull request](https://github.com/CoLRev-Environment/search-query/compare) with a clear description of the changes.
 
 ### Adding a new parser/serializer version
 
