### 2025-09-09 7d03287
revise docs
closes #53
