+++ docs/source/platforms/wos.rst
@@ -5,7 +5,7 @@ Web of Science
 
 Web of Science (WoS) is a multidisciplinary citation database that supports structured search queries with field tags and Boolean operators.
 
-Run a Query
+Run a query
 -----------
 
 Queries in Web of Science can be constructed using:
@@ -19,23 +19,23 @@ Web of Science supports structured queries using field tags such as `TS=` (Topic
 
 .. tip::
 
-   Include field tags explicitly in the `search_string`. Leave the `general_field` empty.
+   Include field tags explicitly in the `search_string`. Leave the general `field` empty.
 
-Store a Query
+Store a query
 -------------
 
 When storing a Web of Science query:
 
 - Use the **Advanced search query string** as the `search_string`.
-- Leave the `general_field` empty unless all terms share the same field tag.
+- Leave the general `field` empty unless all terms share the same field tag.
 
 Example::
 
    (TS="digital health") AND (TS="privacy")
 
 
-List Query Format
----------------------
+List query format
+-----------------
 
 Web of Science allows the construction of **complex queries** by combining previously defined search sets using numbered references (e.g., ``#1 AND #2``). This list-based approach is commonly used in systematic searches where multiple search lines are logically combined.
 
@@ -47,12 +47,12 @@ List queries should be formatted as follows:
 
    {
        "search_string": "1. TS=(digital health OR telemedicine)\n2. TS=(physical activity OR exercise)\n3. #1 AND #2",
-       "general_field": ""
+       "field": ""
    }
 
 Each numbered line represents an individual query component, and later lines can combine previous results using logical operators like ``AND``, ``OR``, or ``NOT``.
 
-Best Practices and Recommendations
+Best practices and recommendations
 ----------------------------------
 
 - Prefer **Advanced Search** for reproducible and structured queries.
