+++ docs/source/index.rst
@@ -94,8 +94,8 @@ The translated query can be saved as follows:
    from search_query import SearchFile
 
    search_file = SearchFile(
-      filename="search-file.json",
-      query_str=wos_query.to_string(platform="wos"),
+      filepath="search-file.json",
+      search_string=wos_query.to_string(),
       platform="wos",
       version="1",
       authors=[{"name": "Tom Brady"}],
@@ -113,8 +113,8 @@ A Jupyter Notebook demo (hosted on Binder) is available here:
 .. image:: https://mybinder.org/badge_logo.svg
    :target: https://mybinder.org/v2/gh/CoLRev-Environment/search-query/HEAD?labpath=docs%2Fsource%2Fdemo.ipynb
 
-Functional Overview
-=======================
+Functional overview
+======================
 
 The search-query package supports the entire lifecycle of academic search query management.
 Below is a high-level overview of the core functionalities:
