### 2025-09-09 b3042ab
docs: use sentence case for headings
