+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-# Change Log
+# Changelog
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
