+++ docs/source/platforms/ebsco.rst
@@ -1,23 +1,23 @@
 .. _ebsco:
 
-EBSCOHost
+EBSCOhost
 =========
 
-EBSCOHost provides access to a wide range of academic databases across disciplines, including business, education, psychology, and health sciences.
+EBSCOhost provides access to a wide range of academic databases across disciplines, including business, education, psychology, and health sciences.
 
-Run a Query
+Run a query
 -----------
 
-EBSCOHost queries can be constructed using the standard `EBSCOHost advanced search interface <https://search.ebscohost.com/>`_ (requires institutional access).
+EBSCOhost queries can be constructed using the standard `EBSCOhost advanced search interface <https://search.ebscohost.com/>`_ (requires institutional access).
 
 When working with `search-query`, extract the **Search terms** from the **Search History** panel or persistent URL for use as the `search_string`.
 
-EBSCOHost query syntax includes field tags such as `AB`, `TI`, `SU`, etc. These should be included directly in the `search_string`.
+EBSCOhost query syntax includes field tags such as `AB`, `TI`, `SU`, etc. These should be included directly in the `search_string`.
 
-Store a Query
+Store a query
 -------------
 
-When storing an EBSCOHost query in a `.json` file or as a string:
+When storing an EBSCOhost query in a `.json` file or as a string:
 
 - Use the **Search History** or the **Search Terms** as the `search_string`.
 
@@ -29,8 +29,8 @@ When storing an EBSCOHost query in a `.json` file or as a string:
 
    Do not use the **persistent link feature**.
 
-List Query Format
---------------------
+List query format
+-----------------
 
 EBSCOhost supports building **multi-line queries** where individual components are defined as numbered search lines and later combined using references like ``S1 AND S2``. This method is often used in advanced or expert search modes.
 
@@ -46,7 +46,7 @@ List queries should be formatted as follows:
 Each numbered item defines a part of the query using EBSCOhost's field syntax. Later lines can combine them using boolean operators like ``AND``, ``OR``, or ``NOT``, allowing for structured, transparent query design.
 
 
-Best Practices and Recommendations
+Best practices and recommendations
 ----------------------------------
 
 - **Use field tags** (e.g., `AB`, `TI`) explicitly in the query string.
