+++ docs/source/query_database/queries/ais_senior_scholars_basket.rst
@@ -26,7 +26,7 @@ Load
     # SO=("European Journal of Information Systems" OR "Information Systems Journal" O...
 
 
-Search string
+Search String
 -------------
 
 .. raw:: html
