### 2025-09-18 5587a62
refactor(search-file): handling of search_history_path and search_results_path
- Replace `filepath` with `search_results_path` in SearchFile
- Add `set_search_results_path()` and `set_search_history_file_path()`
- Add `get_search_history_path()` and use it in `save()`
- Derive history path as <results_dir>/<results_stem>_search_history.json
- Do not serialize `search_history_path` to avoid contradictions
- Improve validation and ensure save directory creation
- Update loader to accept history-path and store it as a private override
