+++ search_query/search_file.py
@@ -6,15 +6,20 @@ import json
 import re
 from pathlib import Path
 from typing import Optional
+from typing import Union
 
 # pylint: disable=too-few-public-methods
 
+_HISTORY_SUFFIX = "_search_history.json"
 
+
+# pylint: disable=too-many-instance-attributes
 class SearchFile:
     """SearchFile class."""
 
+    _search_history_path: Optional[Path] = None
+
     # pylint: disable=too-many-arguments
-    # pylint: disable=too-many-positional-arguments
     def __init__(
         self,
         search_string: str,
@@ -22,7 +27,7 @@ class SearchFile:
         authors: Optional[list[dict]] = None,
         record_info: Optional[dict] = None,
         date: Optional[dict] = None,
-        filepath: Optional[str | Path] = None,
+        search_results_path: Optional[str | Path] = None,
         **kwargs: dict,
     ) -> None:
         self.search_string = search_string
@@ -32,18 +37,62 @@ class SearchFile:
         self.date = date or {}
         # Note: this will be called field_general
         self.field = ""
-        self._filepath = Path(filepath) if filepath else None
+        self.set_search_results_path(search_results_path)
 
         for key, value in kwargs.items():
             setattr(self, key, value)
 
         self._validate_authors(self.to_dict())
 
-    def save(self, filepath: Optional[str | Path] = None) -> None:
-        """Save the search file to a JSON file."""
-        path = Path(filepath) if filepath else self._filepath
+    def set_search_results_path(self, path: Optional[Union[str, Path]]) -> None:
+        """
+        Set or clear the explicit search-results file path.
+        Passing None clears the path.
+        """
         if path is None:
-            raise ValueError("No filepath provided and no previous filepath stored.")
+            self.search_results_path = None
+            return
+
+        self.search_results_path = Path(path)
+        if self.search_results_path.name.endswith(_HISTORY_SUFFIX):
+            raise ValueError(
+                f"search_results_path must not end with '{_HISTORY_SUFFIX}'."
+            )
+
+    def set_search_history_file_path(self, path: Optional[Union[str, Path]]) -> None:
+        """
+        Set or clear the explicit search-history file path (stored privately).
+        Passing None clears the override so the derived path is used again.
+        """
+        if path is None:
+            self._search_history_path = None
+            return
+
+        self._search_history_path = Path(path)
+
+    def get_search_history_path(
+        self, search_history_path: Optional[str | Path] = None
+    ) -> Path:
+        """Get the search history path."""
+
+        if search_history_path:
+            return Path(search_history_path)
+
+        if self._search_history_path is not None:
+            return self._search_history_path
+
+        if getattr(self, "search_results_path", None) is None:
+            raise ValueError(
+                "No search_history_path provided and no search_results_path stored."
+            )
+
+        rp = Path(self.search_results_path)  # type: ignore
+        return rp.with_name(rp.stem + _HISTORY_SUFFIX)
+
+    def save(self, search_history_path: Optional[str | Path] = None) -> None:
+        """Save the search file to a JSON file."""
+        path = self.get_search_history_path(search_history_path)
+        path.parent.mkdir(parents=True, exist_ok=True)
         with open(path, "w", encoding="utf-8") as f:
             json.dump(self.to_dict(), f, indent=4, ensure_ascii=False)
 
@@ -56,6 +105,8 @@ class SearchFile:
             "record_info": self.record_info,
             "date": self.date,
         }
+        if self.search_results_path is not None:
+            data["search_results_path"] = str(self.search_results_path)
         extras = {
             k: v
             for k, v in self.__dict__.items()
@@ -90,9 +141,9 @@ class SearchFile:
[…]
