+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "search-query"
-version = "0.12.0"
+version = "0.13.0"
 description = "Package for managing literature search queries."
 authors = [
     { name = "Peter Eckhardt" },
