### 2025-10-07 #8775 @CBID2
> Hi @CBID2 do you have a date you think you'll complete the review?

Hey @amcandio let's try Friday
