### 2025-11-09 94d4b16
linter_base: extract _handle_unparsed_segment()
