### 2025-11-09 8f90b14
linter_base: check for unbalanced quotes
