+++ search_query/linter_base.py
@@ -74,6 +74,7 @@ class QueryStringLinter:
         # silent: primarily for ListParsers
         self.silent = silent
         self.ignore_failing_linter = ignore_failing_linter
+        self._reported_unbalanced_quotes = False
 
     def add_message(
         self,
@@ -223,12 +224,46 @@ class QueryStringLinter:
         if not segment.strip():
             return
 
+        if self._maybe_report_unbalanced_quotes(start=start, end=end):
+            return
+
         self.add_message(
             QueryErrorCode.TOKENIZING_FAILED,
             positions=[(start, end)],
             details=f"Unparsed segment: '{segment.strip()}'",
             fatal=True,
         )
+
+    def _maybe_report_unbalanced_quotes(self, *, start: int, end: int) -> bool:
+        if self._reported_unbalanced_quotes:
+            return False
+
+        total_quote_count = self.query_str.count('"')
+        if total_quote_count % 2 == 0:
+            return False
+
+        segment = self.query_str[start:end]
+        if segment.count('"') % 2 == 0:
+            prefix = self.query_str[:end]
+            if prefix.count('"') % 2 == 0:
+                return False
+
+        quote_position = self.query_str.rfind('"', 0, end)
+        if quote_position == -1:
+            quote_position = self.query_str.find('"', start, end)
+
+        if quote_position != -1:
+            positions = [(quote_position, quote_position + 1)]
+        else:
+            positions = [(start, end)]
+
+        self.add_message(
+            QueryErrorCode.UNBALANCED_QUOTES,
+            positions=positions,
+            fatal=True,
+        )
+        self._reported_unbalanced_quotes = True
+        return True
 
     def check_general_field(self) -> None:
         """Check the general search field"""
