### 2025-11-09 cba6b3b
cli: highlight successful linting in GREEN
