### 2025-11-09 cdc90cb
linter: select list-parser, rename _run_parser(), clear empty messages
