### 2025-11-09 f41f536
LIST_QUERY_LINE_REGEX: allow for spaces before item-nr
Signed-off-by: Gerit Wagner <gerit.wagner@uni-bamberg.de>
