### 2025-11-09 6a8e69e
revise wos parser
