### 2025-11-09 3c764a7
Warn on escaped Unicode sequences
