### 2025-11-09 01110bb
Warn on escaped Unicode sequences
