### 2025-11-18 9e0ebc1
update cli tests
