+++ test/test_cli.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 """Tests for search query cli"""
 import subprocess
-import sys
 from pathlib import Path
 
 from search_query import load_search_file
@@ -17,9 +16,7 @@ def test_translate_cli() -> None:
 
     result = subprocess.run(
         [
-            sys.executable,
-            "-m",
-            "search_query.cli",
+            "search-query",
             "translate",
             f"--input={input_file}",
             "--to=ebscohost",
@@ -47,13 +44,7 @@ def test_linter_cli() -> None:
     input_file = test_data_dir / "search_history_file_2_linter.json"
 
     result = subprocess.run(
-        [
-            sys.executable,
-            "-m",
-            "search_query.cli",
-            "lint",
-            f"{input_file}",
-        ],
+        ["search-query", "lint", f"{input_file}"],
         capture_output=True,
         text=True,
     )
