### 2025-11-24 #8775 @geritwagner
Hi @crsl4,

It seems possible that @CBID2 may currently not have the availability to complete the review, and we fully understand that unexpected situations can arise — something any review process eventually needs to accommodate.

Could you set a timeline for deciding whether a replacement reviewer is needed? We are happy to facilitate this in any way that helps move the process forward and are flexible with whatever works best for JOSS.

Thanks again for your time and efforts!
