### 2025-11-26 ffea8e9
remove packaging import and dependency
