### 2025-12-11 3bdfde3
deprecate WOS_0005
