### 2025-12-11 8bb48da
deprecate WOS_0005
