### 2025-12-13 bf7cff0
Adjust year value regex
