### 2025-12-16 7fe5986
extend QueryErrorCode.JOURNAL_FILTER_IN_SUBQUERY: cover ISSN etc.
