### 2025-12-20 536eddd
Enforce operator precedence in the parsing step, remove artificial parentheses
