+++ search_query/ebscohost/linter.py
@@ -93,7 +93,7 @@ class EBSCOQueryStringLinter(QueryStringLinter):
         self.check_unknown_token_types()
         self.check_invalid_token_sequences()
         self.check_unbalanced_parentheses()
-        self.add_artificial_parentheses_for_operator_precedence()
+        self._print_unequal_precedence_warning()
         self.check_operator_capitalization()
         self.check_invalid_near_within_operators()
 

[…]
