### 2025-12-21 0638473
Revise build_query_str() to consistently add artificial parentheses when replacing list references with query content
