### 2025-12-21 154f68b
Force numeric comparison in build_query_str()
