### 2025-12-21 9076418
Add test case for different year range notation
