### 2025-12-23 50259fa
Adjust tokenize_operator_node() to interpret gaps between list refs as single token
