### 2025-12-23 14574a6
Add test case for revised tokenize_operator_node() method
