### 2025-12-24 df1e10b
Deprecate WOS_0002
