### 2025-12-24 0dedc37
Add WOS_0013 (Invalid-near-query)
