### 2025-12-24 c7a8d2b
Fix WOS max term amount
