### 2025-12-24 5c08cc1
Remove redundant method overriding
