### 2025-12-25 fd3eb14
Adjust EBSCO tokenizer regex to other platforms to broaden range of parsed terms and simplify patterns
