### 2025-12-27 96beecc
Revise EBSCOHost wildcard validation
