### 2025-12-27 74afe1c
Update fix_ambiguous_tokens() to identify FIELD + OPERATOR token sequences
