### 2025-12-27 11cfd44
Use inline flags in EBSCOHost regex
