### 2025-12-27 430f4f7
Update check_unsupported_wildcards()
