### 2025-12-27 c6bc83d
Call fix_ambiguous_tokens() before combine_subsequent_tokens()
