### 2025-12-28 a76d945
deprecate WOS_0005 (#78)
* deprecate WOS_0005

* update test case

* update message
