+++ docs/source/lint/WOS_0005.rst
@@ -21,4 +21,13 @@ WOS_0005 — year-span-violation
 
 **Typical fix**: The parser automatically adjusts the year span to 5 years.
 
+.. note::
+   **Deprecation Notice**: This linting rule is scheduled for deprecation.
+   Recent evaluations show that the Web of Science (WOS) search interface
+   does **not** enforce a 5-year publication range limit. Queries with
+   broader year spans execute successfully, and applying the publication
+   year filter manually yields identical results. The `WOS documentation <https://webofscience.zendesk.com/hc/en-us/articles/26916347018257-Web-of-Science-Core-Collection-Advanced-Search-Field-Tags>`_
+   mentions the five year limit, but it seems to be a recommendation rather
+   than a strict limit.
+
 **Back to**: :ref:`lint`
