### 2025-12-28 d923a23
docs: drop empty lines
