+++ docs/source/dev_docs/_autosummary/search_query.linter_base.rst
@@ -4,15 +4,6 @@ search\_query.linter\_base
 .. automodule:: search_query.linter_base
 
 
-
-
-
-
-
-
-
-
-
    .. rubric:: Classes
 
    .. autosummary::
