+++ docs/source/dev_docs/_autosummary/search_query.search_file.rst
@@ -4,20 +4,11 @@ search\_query.search\_file
 .. automodule:: search_query.search_file
 
 
-
-
-
-
-
    .. rubric:: Functions
 
    .. autosummary::
 
       load_search_file
-
-
-
-
 
    .. rubric:: Classes
 
