+++ docs/source/dev_docs/_autosummary/search_query.ebscohost.parser.rst
@@ -4,15 +4,6 @@ search\_query.ebscohost.parser
 .. automodule:: search_query.ebscohost.parser
 
 
-
-
-
-
-
-
-
-
-
    .. rubric:: Classes
 
    .. autosummary::
