+++ docs/source/dev_docs/_autosummary/search_query.upgrade.rst
@@ -4,11 +4,6 @@ search\_query.upgrade
 .. automodule:: search_query.upgrade
 
 
-
-
-
-
-
    .. rubric:: Functions
 
    .. autosummary::
