+++ docs/source/dev_docs/_autosummary/search_query.ebscohost.linter.rst
@@ -4,15 +4,6 @@ search\_query.ebscohost.linter
 .. automodule:: search_query.ebscohost.linter
 
 
-
-
-
-
-
-
-
-
-
    .. rubric:: Classes
 
    .. autosummary::
