+++ docs/source/dev_docs/_autosummary/search_query.generic.v_1.rst
@@ -4,27 +4,10 @@ search\_query.generic.v\_1
 .. automodule:: search_query.generic.v_1
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 .. rubric:: Modules
 
 .. autosummary::
    :toctree:
    :recursive:
 
-   search_query.generic.v_1.serializer
+   serializer
