+++ docs/source/dev_docs/_autosummary/search_query.query_or.rst
@@ -4,15 +4,6 @@ search\_query.query\_or
 .. automodule:: search_query.query_or
 
 
-
-
-
-
-
-
-
-
-
    .. rubric:: Classes
 
    .. autosummary::
