+++ docs/source/dev_docs/_autosummary/search_query.query_and.rst
@@ -4,15 +4,6 @@ search\_query.query\_and
 .. automodule:: search_query.query_and
 
 
-
-
-
-
-
-
-
-
-
    .. rubric:: Classes
 
    .. autosummary::
