+++ docs/source/dev_docs/_autosummary/search_query.serializer_base.rst
@@ -4,15 +4,6 @@ search\_query.serializer\_base
 .. automodule:: search_query.serializer_base
 
 
-
-
-
-
-
-
-
-
-
    .. rubric:: Classes
 
    .. autosummary::
