+++ docs/source/dev_docs/_autosummary/search_query.constants.rst
@@ -4,15 +4,6 @@ search\_query.constants
 .. automodule:: search_query.constants
 
 
-
-
-
-
-
-
-
-
-
    .. rubric:: Classes
 
    .. autosummary::
