+++ docs/source/dev_docs/_autosummary/search_query.wos.v_1.rst
@@ -4,29 +4,12 @@ search\_query.wos.v\_1
 .. automodule:: search_query.wos.v_1
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 .. rubric:: Modules
 
 .. autosummary::
    :toctree:
    :recursive:
 
-   search_query.wos.v_1.parser
-   search_query.wos.v_1.serializer
-   search_query.wos.v_1.translator
+   parser
+   serializer
+   translator
