### 2025-12-28 052ce31
assign_linter_messages: use object linter instead of parameter
