+++ search_query/ebscohost/parser.py
@@ -477,7 +477,7 @@ class EBSCOListParser(QueryListParser):
         except QuerySyntaxError as exc:
             raise exc
         finally:
-            self.assign_linter_messages(query_parser.linter.messages, self.linter)
+            self.assign_linter_messages(query_parser.linter.messages)
 
             self.linter.check_status()
 
