+++ search_query/pubmed/parser.py
@@ -343,7 +343,7 @@ class PubmedListParser(QueryListParser):
         except QuerySyntaxError as exc:
             raise exc
         finally:
-            self.assign_linter_messages(query_parser.linter.messages, self.linter)
+            self.assign_linter_messages(query_parser.linter.messages)
 
             self.linter.check_status()
 
