+++ search_query/wos/parser.py
@@ -432,7 +432,7 @@ class WOSListParser(QueryListParser):
         except QuerySyntaxError as exc:
             raise exc
         finally:
-            self.assign_linter_messages(query_parser.linter.messages, self.linter)
+            self.assign_linter_messages(query_parser.linter.messages)
             self.linter.check_status()
 
         query.set_platform_unchecked(PLATFORM.WOS.value, silent=True)
