### 2025-12-28 a0dcf7c
remove duplicate code
