### 2025-12-29 a34388a
Auto-update documentation indices
