### 2026-01-07 3800919
Use parser_base method for combining terms in EBSCO
