### 2026-01-07 ee75535
Adjust Ebsco term regex to exclude quotes
