### 2026-01-07 c49db82
Revise combine_subsequent_terms
