### 2026-01-07 f0ec297
Add missing operator validation step in WOS
