### 2026-01-08 16b034d
Handle invalid characters during token validation
