### 2026-01-16 12c0c4c
Revise quoted search term tokenization and validation
