+++ search_query/constants.py
@@ -33,6 +33,7 @@ class TokenTypes(Enum):
     PROXIMITY_OPERATOR = "PROXIMITY_OPERATOR"
     FIELD = "FIELD"
     TERM = "TERM"
+    QUOTATION_MARK = "QUOTATION_MARK"
     PARENTHESIS_OPEN = "PARENTHESIS_OPEN"
     PARENTHESIS_CLOSED = "PARENTHESIS_CLOSED"
     RANGE_OPERATOR = "RANGE_OPERATOR"

[…]
