### 2026-01-19 bfa0bd5
Revise WOS general field validation, improvements to invalid token sequence validation
