### 2026-01-19 c127f22
Revise _print_unequal_precedence_warning()
