### 2026-01-20 7b05717
Add document type field for WOS
