+++ search_query/wos/constants.py
@@ -22,6 +22,7 @@ SYNTAX_GENERIC_MAP = {
         Fields.KEYWORDS_PLUS,
     },
     "LA=": {Fields.LANGUAGE},
+    "DT=": {Fields.PUBLICATION_TYPE},
     "PY=": {Fields.YEAR_PUBLICATION},
     "AD=": {Fields.ADDRESS},
     "AI=": {Fields.AUTHOR_IDENTIFIERS},
@@ -66,6 +67,7 @@ DOI_FIELD_REGEX: re.Pattern = re.compile(r"do=|doi=", re.IGNORECASE)
 _RAW_PREPROCESSING_MAP = {
     "AB=": r"ab=|abstract=",
     "LA=": r"la=|language=",
+    "DT=": r"dt=|document type=",
     "AD=": r"ad=|address=",
     "ALL=": r"all=|all fields=",
     "AI=": r"ai=|author identifiers=",
