### 2026-01-20 89468b0
Update WOS tests
