### 2026-01-22 256946e
Normalize search fields for query tree validation
