### 2026-01-23 e95d9f2
Revise potential wildcard validation
