### 2026-01-23 88dfffd
Minor changes to fatal error handling
