### 2026-01-23 aac8b6d
Fix PubMed false invalid-proximity-use error case
