### 2026-01-24 155554f
WOS: Treat non-search string = as invalid character
