### 2026-01-24 95c54b0
Format platform str in unsupported-field details
