+++ test/ebscohost/test_ebscohost_parser_v_1.py
@@ -246,7 +246,7 @@ def test_invalid_token_sequences(
                     "message": "Search field is not supported for this database",
                     "is_fatal": True,
                     "position": [(0, 2)],
-                    "details": "Search field AI at position (0, 2) is not supported. Supported fields for PLATFORM.EBSCO: TI|AB|TP|TX|AU|SU|SO|IS|IB|LA|KW|DE|MH|MM|MW|ZY|ZU|PT|ZT|DT|XB",
+                    "details": "Search field AI at position (0, 2) is not supported. Supported fields for EBSCOHOST: TI|AB|TP|TX|AU|SU|SO|IS|IB|LA|KW|DE|MH|MM|MW|ZY|ZU|PT|ZT|DT|XB",
                 }
             ],
         ),
@@ -259,7 +259,7 @@ def test_invalid_token_sequences(
                     "message": "Search field is not supported for this database",
                     "is_fatal": True,
                     "position": [(0, 2)],
-                    "details": "Search field AI at position (0, 2) is not supported. Supported fields for PLATFORM.EBSCO: TI|AB|TP|TX|AU|SU|SO|IS|IB|LA|KW|DE|MH|MM|MW|ZY|ZU|PT|ZT|DT|XB",
+                    "details": "Search field AI at position (0, 2) is not supported. Supported fields for EBSCOHOST: TI|AB|TP|TX|AU|SU|SO|IS|IB|LA|KW|DE|MH|MM|MW|ZY|ZU|PT|ZT|DT|XB",
                 }
             ],
         ),
