### 2026-01-24 #8775 @tnehezd
Hey, @geritwagner!

I went through most of your examples, and I've finished the paper and the checklist. Very nice work! :)

I have opened two issues:
https://github.com/CoLRev-Environment/search-query/issues/80
https://github.com/CoLRev-Environment/search-query/issues/81

And I don't know if it's on my side, or something else, but copying the example from this site gives me an error, also:

https://colrev-environment.github.io/search-query/automate.html

```
    from colrev.packages.crossref.src import crossref_api
ModuleNotFoundError: No module named 'colrev'
```

Otherwise, it's a handy package, with an impressive doc site.

Best regards,
Dóri
