### 2026-01-25 138cdd9
docs: add installation instructions for automate example
