### 2026-01-25 #8775 @geritwagner
Hi @tnehezd,

thanks a lot for taking the time to go through the examples, and for the kind words — much appreciated! 🙂 The students involved in this project will be very happy when I share your feedback with them.

Regarding the missing colrev module: search-query itself does not handle record retrieval. Implementing retrieval, parsing, and saving directly from APIs like Crossref would require a fairly long example, so we decided to rely on colrev for that part (it already covers retrieval, parsing, and writing). To make this clearer, I’ve added explicit instructions to install colrev on the automate page:
https://colrev-environment.github.io/search-query/automate.html

Hope these revisions address your point.

I’ve just fixed the second issue — thanks for flagging it.

For the first issue, I will revise the Binder example and also to provide a slightly more comprehensive demo of the package. I will send an update when everything is ready. Thanks again for the helpful feedback!

Best,
Gerit
