### 2026-01-25 1cf9a33
Revise WOS check_missing_fields()
