### 2026-01-26 #8775 @tnehezd
Hey @geritwagner,

I've closed the issue about the demo! Great work, seems fine!

Thanks for the upgraded documentation about Crossref. However, it seems that CrossrefApi does not support URL keyword:

https://github.com/CoLRev-Environment/search-query/issues/82

Best regards,
Dóri
