### 2026-01-27 ee8413c
Update invalid char highlighting
