### 2026-01-27 e47b8cb
wos: fix translator (failing tests)
