### 2026-01-28 59a96b9
refine docs and code comments/format
