+++ search_query/wos/translator.py
@@ -3,8 +3,9 @@
 from __future__ import annotations
 
 import re
-from search_query.constants import PLATFORM
+
 from search_query.constants import Fields
+from search_query.constants import PLATFORM
 from search_query.query import Query
 from search_query.query import SearchField
 from search_query.query_or import OrQuery
