+++ CITATION.cff
@@ -14,6 +14,6 @@ authors:
 - family-names: "Wagner"
   given-names: "Gerit"
 title: "Search-query: A Python package for queries in academic literature searches."
-version: 0.12.0
-date-released: 2025-09-18
+version: 0.14.0
+date-released: 2026-01-28
 url: "https://github.com/ColRev-Environment/search-query"
