### 2026-01-28 aa88c35
docs/automate: pin recent colrev version and use simpler example
