### 2026-01-28 75386b5
docs/automate: pin colrev to 0.16.1 (jupyter compatibility)
