### 2026-01-28 #8775 @tnehezd
Hi @geritwagner!

I've checked the modified documentation with the new colrev version, and everything works fine! However, as I mentioned in the opened (and then closed) issue, I suggest mentioning in the documentation that colrev 0.16.1 requires python3.10 or above.

Besides, I'm finished the review and pleased to say I accept the manuscript and the package for publication. :)

@crsl4, please let me know if I have anything more with this process.

Best regards,
Dóri
