### 2026-01-28 5135408
docs/auto: add note on Python3.10+ requirement (#82)
