### 2026-01-28 #8775 @geritwagner
Hi @tnehezd,

many thanks for the constructive suggestions and the very timely responses — this is much appreciated by the whole team.

I have just pushed the changes related to the Python ≥3.10 requirement. Thanks you for recommending the acceptance of our work — this is great news!

Best regards,
Gerit
