### 2026-01-29 d98a610
docs: update functional overview
