### 2026-01-29 a5261c1
add dist-smoke-test
