+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'search-query: An Open-Source Python Library for Academic Search Queries'
+title: 'search-query: An open source Python library for academic search queries'
 tags:
 - Python
 - Search query
@@ -40,13 +40,13 @@ header-includes: |
 
 # Summary
 
-*search-query* is an open-source Python library designed to load, lint, translate, save, improve, and automate academic literature search queries.
-Unlike existing proprietary and web-based tools, *search-query* offers programmatic access and thereby supports integration into research workflows, contributing to the automation of systematic literature reviews.
-It currently provides query translation between Web of Science, PubMed, and EBSCOHost, while its validation features detect syntactical errors and inconsistencies that may compromise queries.
-The parsers for these three databases are tested against a comprehensive set of peer-reviewed queries from searchRxiv.
-By addressing key gaps in literature search tools, *search-query* contributes to the correctness and efficiency of systematic reviews and meta-analyses.
+*search-query* is an open source Python library for academic literature search queries.
+Its core capabilities include linting to detect syntactic and logical errors, translation to convert queries across database-specific syntaxes, improvement to programmatically refine queries, and automation to integrate searches into reproducible workflows.
+The library currently supports *Web of Science*, *PubMed*, and *EBSCOHost*, with parsers validated against a comprehensive set of peer-reviewed queries from *searchRxiv*.
+Unlike existing proprietary and web-based software, *search-query* offers programmatic access and thereby supports integration into research workflows, contributing to the automation of systematic literature reviews.
+<!-- By addressing key gaps in literature search tools, *search-query* contributes to the correctness and efficiency of systematic reviews and meta-analyses. -->
 
-Keywords: Python, search query, literature search, literature review.
+Keywords: Python, search query, literature search, literature review, meta-analysis.
 
 <!-- 
 - entry: enter queries from cli/programmatically
@@ -67,7 +67,7 @@ Planned:
 - effectiveness: estimate retrieval performance based on a given sample
 - sample extension: strategically extend the sample to improve accuracy of retrieval performance estimates
 
-Properties: tested, extensible, open-source
+Properties: tested, extensible, open source
 
 - A substantial percentage of peer-reviewed literature search queries has syntactical errors, ambiguities (??), and inconsistencies
 - Existing tools do not sufficiently validate queries (often, it is unclear whether they were actually used)
@@ -84,50 +84,48 @@ https://link.springer.com/referenceworkentry/10.1007/978-0-387-39940-9_1070
 # Statement of Need
 
 Researchers conducting meta-analyses and other types of literature reviews rely on database searches as the primary search technique [@Hiebl2023].
-This primarily includes the design of Boolean queries, and the execution in academic databases.
+This involves the design of Boolean queries, the translation for specific databases, and the execution.
 Requirements include:
 (1) query translation, i.e., parsing a query in one database specific syntax and serializing it in another syntax [e.g., @AlZubidyCarver2019; @SturmSunyaev2019],
-(2) query validation, i.e., identifying syntactical errors or warning of known database errors [e.g., @LiRainer2023; @SinghSingh2017],
+(2) query linting, i.e., identifying syntactical errors or warning of known database errors [e.g., @LiRainer2023; @SinghSingh2017],
 (3) query improvement, i.e., manipulation of query objects to understand and enhance performance [e.g., @ScellsZucconKoopman2020], and
-(4) integration capability, i.e., offering programmatic or API access to integrate in a tool pipeline [e.g., @BellerClarkTsafnatEtAl2018; @OConnorTsafnatGilbertEtAl2018].
[…]
