### 2026-01-29 b08cdc4
Revise check_redundant_terms()
