### 2026-01-30 846207f
make SearchFile(query=...) default
rationale: passing the query object can prevent
mismatches, which occur when users pass the
query string and platform separately.
