+++ docs/source/index.rst
@@ -94,9 +94,7 @@ The translated query can be saved as follows:
    from search_query import SearchFile
 
    search_file = SearchFile(
-      search_string=wos_query.to_string(),
-      platform="wos",
-      version="1",
+      query=wos_query,
       authors=[{"name": "Tom Brady"}],
       record_info={},
       date={}
