### 2026-01-30 e42359d
update SearchFile() to more robust version
