### 2026-01-30 #8775 @geritwagner
Hi @crsl4 ,

just a quick update: I have just read and polished the whole manuscript, gone through all documentation examples, and re-ran the demo notebook — everything works as expected, and I am very happy with the current version. From our side, it should be ready.

We also received two acceptance recommendations from the reviewers. Please let us know if there’s anything else we should do to help move this toward publication.

Many thanks,
Gerit
