+++ CITATION.cff
@@ -14,6 +14,6 @@ authors:
 - family-names: "Wagner"
   given-names: "Gerit"
 title: "Search-query: A Python package for queries in academic literature searches."
-version: 0.14.1
-date-released: 2026-01-28
+version: 0.15.0
+date-released: 2026-02-06
 url: "https://github.com/ColRev-Environment/search-query"
