### 2026-02-06 #8775 @geritwagner
I completed the author tasks from the checklist:

The version is `0.15.0`.

The doi is `10.5281/zenodo.18502307`
