### 2026-02-11 #8775 @geritwagner
@crsl4 , I double-checked the suggested DOI. It refers to a review of the Yale MeSH Analyzer rather than to the tool itself. Using that DOI would therefore credit the review authors instead of the original developers.

Also, the Yale MeSH Analyzer is not cited in the manuscript text; it only appears in the BibTeX file and in commented sections. Substantively, it supports MeSH term analysis rather than directly assisting in the development or validation of search queries as discussed in the paper.

Please let me know whether we should remove the entry from the BibTeX file or whether you prefer to leave it as is.
