### 2026-02-14 #8775 @crsl4
So, it is not a reference in the manuscript text? If so, then maybe removing from the bibtex would make more sense so that it is not flagged.
