+++ paper.bib
@@ -40,21 +40,6 @@
    url                           = {https://link.springer.com/article/10.1007/s12599-018-0569-6},
 }
 
-
-@online{GrossettaNardiniWang2023,
-   author                        = {Grossetta Nardini, Holly and Wang, Lei},
-   title                         = {Yale MeSH Analyzer},
-   year                          = {2023},
-   url                           = {https://mesh.med.yale.edu/},
-   cost                          = {Free},
-   description                   = {MeSH analysis tool to help identify problems in the search strategy.},
-   language                      = {eng},
-   note                          = {latest version 2023, https://mesh.med.yale.edu/help#how-to-cite, MeSH analysis tool to help identify problems in the search strategy.},
-   related_publications          = {n/a},
-   review_family                 = {Systematic, Rapid, Qualitative, Scoping, Mapping, Mixed Methods, Reviews of reviews, Other, Multiple},
-   supports                      = {Search},
-}
-
 @online{ClarkGlasziouDelMarEtAl2020a,
    author                        = {Clark, Justin Michael and Glasziou, Paul and Del Mar, Chris and Bannach-Brown, Alexandra and Stehlik, Paulina and Scott, Anna},
    title                         = {Polyglot Search},
