### 2026-02-16 #8775 @editorialbot
:wave: @openjournals/misc-eics, this paper is ready to be accepted and published.
