### 2026-02-20 #8775 @geritwagner
Thank you, @arfon and @crsl4, for handling the submission so smoothly! And sincere thanks to our reviewers @tnehezd and @amcandio for the constructive and thoughtful feedback — it truly helped improve both the package and the paper.

Very proud of this team effort — @Peteer98, @katharinaernst, @annaglr, @ThomasFleischmann, @k-schnickmann, @LaureenTh, and everyone involved in bringing this from the theses to a published JOSS paper. 🚀
